% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_scores.R
\name{act_dims}
\alias{act_dims}
\title{Active Dimension (Not validated. Might be buggy)}
\usage{
act_dims(C, X, y, k = ncol(C), alpha = 0.05, all_sets = TRUE, verbose = TRUE)
}
\arguments{
\item{C}{Constantines C matrix (e.g. from C_bass, C_mc, or C_gp)}

\item{X}{the original input variables}

\item{y}{the original response variable (mod$y when using C_bass(mod))}

\item{k}{The maximum number of columns of W to consider}

\item{alpha}{significance threshold for testing procedure}

\item{all_sets}{should all dimension sets be returned? Or just the smallest set.}

\item{verbose}{should progress be printed}
}
\value{
a list of active subspace dimensions
}
\description{
This function estimates the dimensions of the active subspace using a sequential testing approach
}
