% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_scores.R
\name{act_scores}
\alias{act_scores}
\title{Activity Scores}
\usage{
act_scores(C, k = 1, plt = FALSE, norm = FALSE)
}
\arguments{
\item{C}{Constantines C matrix (e.g. from C_bass, C_mc, or C_gp)}

\item{k}{The number of columns of W to consider}

\item{plt}{Logical, should a plot be made?}

\item{norm}{Logical, should activity scores be normalized to have a maximum value of 1?}
}
\value{
the activity scores
}
\description{
This function computes the activity scores for main effects of the variables
}
