% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bassfunc2bass.R
\name{bassfunc2bass}
\alias{bassfunc2bass}
\title{Convert functional BASS model to BASS model}
\usage{
bassfunc2bass(bfm)
}
\arguments{
\item{bfm}{an object of class bass, where a functional variable has been specified.}
}
\description{
The argument to this function is the output of a \code{bass()} call when a single functional variable is specified using the \code{xx.func} argument.
Note that the resulting model may not be a valid bass object for some applications,
but the resulting model can be passed to \code{concordance::C_bass()} and related functions.
}
\examples{
#The following are equivalent
n <- 100 #Number of observations
p <- 4   #Number of variables (beyond p = 2, variables are inert)
X <- matrix(runif(n*p), nrow=n)
y <- apply(X, 1, ff1)
gm <- gbass(X, Y, nmcmc=1000, nburn=901)
bm <- gm2bm(gm)
sob <- sobol(bm)
plot(sob)
}
