% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bea2002_desc}
\alias{bea2002_desc}
\title{BEA Description for year 2002}
\format{
A data frame with 430 rows and 2 variables:
\describe{
  \item{code}{BEA Code}
  \item{desc}{BEA Description}
}
}
\source{
\url{https://www.bea.gov/industry/benchmark-input-output-data}
}
\usage{
bea2002_desc
}
\description{
A dataset containing the description of products under the BEA industry classification in 2002
}
\keyword{datasets}
