% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bea2002_naics2002}
\alias{bea2002_naics2002}
\title{BEA2002-NAICS2002 Concordance}
\format{
A data frame with 1,335 rows and 4 variables:
\describe{
  \item{BEA2002}{6-digit BEA2002 Code}
  \item{NAICS2002_6d}{6-digit NAICS2002 Code}
  \item{NAICS2002_4d}{4-digit NAICS2002 Code}
  \item{NAICS2002_2d}{2-digit NAICS2002 Code}
}
}
\source{
\url{https://www.bea.gov/industry/benchmark-input-output-data}
}
\usage{
bea2002_naics2002
}
\description{
A dataset containing concordances between BEA2002 and NAICS2002 classification.
}
\keyword{datasets}
