% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bea2012_desc}
\alias{bea2012_desc}
\title{BEA Description for year 2012}
\format{
A data frame with 409 rows and 2 variables:
\describe{
  \item{code}{BEA Code}
  \item{desc}{BEA Description}
}
}
\source{
\url{https://apps.bea.gov/iTable/index_industry_io.cfm}
}
\usage{
bea2012_desc
}
\description{
A dataset containing the description of products under the BEA industry classification in 2007 and 2012
}
\keyword{datasets}
