% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bea2012_naics2012}
\alias{bea2012_naics2012}
\title{BEA2012-NAICS2012 Concordance}
\format{
A data frame with 1,376 rows and 4 variables:
\describe{
  \item{BEA2012}{6-digit BEA2012 Code}
  \item{NAICS2012_6d}{6-digit NAICS2012 Code}
  \item{NAICS2012_4d}{4-digit NAICS2012 Code}
  \item{NAICS2012_2d}{2-digit NAICS2012 Code}
}
}
\source{
\url{https://apps.bea.gov/iTable/index_industry_io.cfm}
}
\usage{
bea2012_naics2012
}
\description{
A dataset containing concordances between BEA2012 and NAICS2012 classification.
}
\keyword{datasets}
