% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bec4_desc}
\alias{bec4_desc}
\title{BEC4 Description}
\format{
A data frame with 50 rows and 2 variables:
\describe{
  \item{code}{BEC4 Code}
  \item{desc}{BEC4 Description}
}
}
\source{
\url{https://comtrade.un.org/data/cache/classificationBEC.json}
}
\usage{
bec4_desc
}
\description{
A dataset containing the description of products under the BEC4
classification.
}
\keyword{datasets}
