% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bec_desc}
\alias{bec_desc}
\title{BEC Description}
\format{
A data frame with 29 rows and 2 variables:
\describe{
  \item{code}{BEC Code}
  \item{desc}{BEC Description}
}
}
\source{
\url{https://comtrade.un.org/data/cache/classificationBEC.json}
}
\usage{
bec_desc
}
\description{
A dataset containing the description of products under the BEC
classification.
}
\keyword{datasets}
