% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{borehole_grad}
\alias{borehole_grad}
\title{The Gradient of Borehole Function}
\usage{
borehole_grad(xx, design = 0.5, adjust = TRUE)
}
\arguments{
\item{xx}{the 7 inputs, restricted to the unit interval}

\item{design}{the radius of the borehole (typically rw)}

\item{adjust}{logical. adjustment for scaling needed?}
}
\value{
The output of the borehole function
}
\description{
This function returns the gradient of the borehole function.
}
\details{
PARAMETER RANGES
rw in \verb{[0.05, 0.15]} 	radius of borehole (m)
r in \verb{[100, 50000]} 	radius of influence (m)
Tu in \verb{[63070, 115600]}    	transmissivity of upper aquifer (m2/yr)
Hu in \verb{[990, 1110]} 	potentiometric head of upper aquifer (m)
Tl in \verb{[63.1, 116]} 	transmissivity of lower aquifer (m2/yr)
Hl in \verb{[700, 820]} 	potentiometric head of lower aquifer (m)
L in \verb{[1120, 1680]} 	length of borehole (m)
Kw in \verb{[9855, 12045]} 	hydraulic conductivity of borehole (m/yr)
}
