% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_prior.R
\name{build_prior}
\alias{build_prior}
\title{Build Prior Method for C_bass and Cfg_bass}
\usage{
build_prior(
  dist,
  trunc = NULL,
  mean = NULL,
  sd = NULL,
  shape1 = NULL,
  shape2 = NULL,
  shape = NULL,
  scale = NULL
)
}
\arguments{
\item{dist}{A vector of length p. Valid entries include "uniform", "normal", "beta", "gamma".}

\item{trunc}{A matrix of dimension px2 (rows are recycled if nrow < p). \code{Inf} is a valid entry.}

\item{mean}{A p-vector of means (used for normal only)}

\item{sd}{A p-vector of sds (used for normal only)}

\item{shape1}{A p-vector of shape1 parameters for beta prior}

\item{shape2}{A p-vector of shape2 parameters for beta prior}

\item{shape}{A p-vector of shape parameters for gamma prior}

\item{scale}{A p-vector of scale parameters for gamma prior}
}
\value{
a list which can be passed into C_bass or Cfg_bass as a prior.
}
\description{
A quick way to build priors for use in \code{C_bass} and \code{Cfg_bass}.
For more complicated priors, such as mixture distributions, see details in \code{?C_bass}.
}
\details{
All vectors and matrix rows are recycled for parameters. The vector \code{dist} cannot be recyled as it defines p.
}
