\name{BWsigma}
\alias{sigma_data}

\docType{data}
\title{Table of trade elasticities}

\description{Trade elasticity look-up table, using HS codes only}

\usage{data("BWsigma")}

\format{
  A data frame with 13972 observations on the following 4 variables.
  \itemize{
    \item{\code{hs}}
    \item{\code{sigma}}
    \item{\code{hs4}}
    \item{\code{hs6}}
  }
}

\source{Data from Broda and Weinstein's site 
http://www.columbia.edu/~dew35/TradeElasticities/TradeElasticities.html

See also C. Broda and D. Weinstein, "Globalization and the Gains from Variety," 	
 Quarterly Journal of Economics Volume 121, Issue 2 - May 2006}

\keyword{ datasets, sigma }