% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_Matrix.R
\name{C_bass}
\alias{C_bass}
\title{Estimate the Constantine Matrix with BASS}
\usage{
C_bass(mod, prior = NULL, mcmc.use = NULL, scale01 = FALSE)
}
\arguments{
\item{mod}{a fitted BASS model}

\item{prior}{NULL (default) (0,1])prior for each variable. See details for required structure of prior}

\item{mcmc.use}{set of indices telling which mcmc draws to use}

\item{scale01}{logical (default FALSE). When TRUE, the the C matix corresponds to the (0, 1)-scaled inputs rather than the original inputs.}
}
\value{
A list representing the posterior distribution of the Constantine matrix.
}
\description{
Closed form estimator of the C matrix using a BASS model
}
\details{
prior should be a list of length p (one object for each variable). Each element of prior should be a named list with fields.
See also the concordance::build_prior() function.
\itemize{
\item dist - ("uniform", "normal", "beta", "gamma").
\item trunc - truncation bounds (a, b). These should be c(0, 1) for "beta" and c(0, Inf) for "gamma".
\item mean - vector of means (mixture of normals only)
\item sd - vector of sds (mixture of normals only)
\item shape1, shape2 - shape parameters for beta distribution
\item shape, scale - parameters for gamma distribution
\item weights - vector of mixture weights (currently only compatible with dist="normal")
}
}
