% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_pca.R
\name{C_bassPCA}
\alias{C_bassPCA}
\title{Estimate C matrix with bassPCA as a function of t}
\usage{
C_bassPCA(modPCA, prior = NULL, mcmc.use = NULL, func.use = NULL)
}
\arguments{
\item{modPCA}{a fitted model of class bassBasis from bassPCA function}

\item{prior}{NULL (default) \verb{[0, 1]} prior for each variable. See details for required structure of prior}

\item{mcmc.use}{vector of mcmc indices to be used for both models. Otherwise, a matrix}

\item{func.use}{a vector of points of the functional variable to use}
}
\value{
A list returning the (posterior samples?) of the C matrix for each point specified in func.use
}
\description{
Closed form estimator of the C(t) matrix using a BASS model
}
\details{
This function works by converting the linear combination of bass models to a single bass model. See C_bass for more details
}
