% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_pca.R
\name{C_bassPCA_v2}
\alias{C_bassPCA_v2}
\title{Estimate C matrix with bassPCA as a function of t}
\usage{
C_bassPCA_v2(modPCA, prior = NULL, mcmc.use = NULL, func.use = NULL)
}
\arguments{
\item{modPCA}{a fitted model of class bassBasis from bassPCA function}

\item{prior}{NULL (default) \verb{[0, 1]} prior for each variable. See details for required structure of prior}

\item{mcmc.use}{vector of mcmc indices to be used for both models. Otherwise, a matrix}

\item{func.use}{a vector of points of the functional variable to use}
}
\value{
A list returning the (posterior samples?) of the C matrix for each point specified in func.use
}
\description{
Closed form estimator of the C(t) matrix using a BASS model.
An alternative approach see details.
}
\details{
This function works by decomposing the C of a linear combination into the pairwise Cfg matrices of the components. See C_bass for more details
}
