% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_Matrix.R
\name{C_mc}
\alias{C_mc}
\title{C matrix with Monte Carlo}
\usage{
C_mc(f, measure, grad = FALSE, nmc = 10000, seed = NULL, ...)
}
\arguments{
\item{f}{the function f (or gradient of f, if grad=TRUE)}

\item{measure}{the number of inputs in f. See details for more sophisticated use (for non-uniform measure)}

\item{grad}{if TRUE f is assumed to return the gradient of f. When FALSE, forward diff is used for gradient approximation.}

\item{nmc}{the number of Monte Carlo replications}

\item{seed}{optional. seed for MC draws}

\item{...}{additional arguments passed to f()}
}
\value{
the approximated C matrix
}
\description{
Approximates Constantine's C with Monte Carlo for a function f
}
\details{
\code{measure} should be an argument-free function which simulates a draw x ~ p(x) where p is the prior measure. If \code{measure} is numeric, then Monte Carlo draws are simulated from the standard uniform distribution as \code{runif(measure[1])}.
}
