% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_pca.R
\name{Cfg_bassPCA_v2}
\alias{Cfg_bassPCA_v2}
\title{Estimate Cfg matrix with bassPCA as a function of t}
\usage{
Cfg_bassPCA_v2(
  modPCA1,
  modPCA2,
  prior = NULL,
  mcmc.use = NULL,
  func.use = NULL
)
}
\arguments{
\item{modPCA1}{a fitted model of class bassBasis from bassPCA function}

\item{modPCA2}{a fitted model of class bassBasis from bassPCA function}

\item{prior}{NULL (default) \verb{[0, 1]} prior for each variable. See details for required structure of prior}

\item{mcmc.use}{vector of mcmc indices to be used for both models. Otherwise, a matrix}

\item{func.use}{a vector of points of the functional variable to use}
}
\value{
A list returning the (posterior samples?) of the Cfg matrix for each point specified in func.use
}
\description{
Closed form estimator of the Cfg(t) matrix using a BASS model.
An alternative approach, see details.
}
\details{
This function works by decomposing the Cfg of a linear combination into the pairwise Cfigj matrices of the components. See Cfg_bass for more details
}
