% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_scores.R
\name{coact_scores}
\alias{coact_scores}
\title{Co-Activity Scores}
\usage{
coact_scores(V, q = 1, signed = TRUE, plt = FALSE, norm = FALSE)
}
\arguments{
\item{V}{The symmetrized co-Constantine matrix from Cfg_bass()}

\item{q}{The number of columns of W to consider}

\item{signed}{Use signed or unsigned version?}

\item{plt}{Logical, should a plot be made?}

\item{norm}{Logical, should activity scores be normalized to have a length of 1?}
}
\value{
the coactivityactivity scores
}
\description{
This function computes the activity scores for main effects of the variables
}
