% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance_r.R
\name{coactive_bass}
\alias{coactive_bass}
\title{Concordance Analysis (with BASS)}
\usage{
coactive_bass(mod1, mod2, prior = NULL, mcmc.use = NULL, q = 1, ...)
}
\arguments{
\item{mod1}{BASS model representing function 1}

\item{mod2}{BASS model representing function 2}

\item{prior}{NULL (default) Uniform(0,1) prior for each variable. See details fr required prior structure.}

\item{mcmc.use}{a vector of mcmc replications to use. Can also be a 2-column matrix with indices for f1 and f2.}

\item{q}{order for the activity score measures}

\item{...}{additional arguments passed fd_grad()}
}
\value{
Estimates of C1, C2, C12, V12, conc(f1, f2), contributions and coactivity scores
}
\description{
Computes the concordance between mod1 and mod2 (BASS models representing f1 and f2)
}
\details{
\code{measure} should be an argument-free function which simulates a draw x ~ p(x) where p is the prior measure. If \code{measure} is numeric, then Monte Carlo draws are simulated from the standard uniform distribution as \code{runif(measure[1])}.
}
