% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance_r.R
\name{conc}
\alias{conc}
\title{Concordance}
\usage{
conc(f, g, measure, grad = FALSE, nmc = 10000, ...)
}
\arguments{
\item{f}{the function f (or gradient of f, if grad=TRUE)}

\item{g}{the function g (or gradient of g, if grad=TRUE)}

\item{measure}{the number of inputs in f and g. See details for more sophisticated use (for non-uniform measure)}

\item{grad}{if TRUE f and g are assumed to return gradients. When FALSE, forward diff is used for approximation.}

\item{nmc}{the number of monte carlo replications}

\item{...}{additional arguments passed fd_grad()}
}
\value{
the concordance between functions f and g
}
\description{
Computes the concordance between f and g
}
\details{
\code{measure} should be an argument-free function which simulates a draw x ~ p(x) where p is the prior measure. If \code{measure} is numeric, then Monte Carlo draws are simulated from the standard uniform distribution as \code{runif(measure[1])}.
}
