% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance_r.R
\name{conc_analysis}
\alias{conc_analysis}
\title{Concordance Analysis}
\usage{
conc_analysis(
  f,
  g,
  measure,
  grad = FALSE,
  nmc = 10000,
  names = c("f", "g"),
  seed = NULL,
  ...
)
}
\arguments{
\item{f}{the function f (or gradient of f, if grad=TRUE)}

\item{g}{the function g (or gradient of g, if grad=TRUE)}

\item{measure}{the number of inputs in f and g. See details for more sophisticated use (for non-uniform  measure)}

\item{grad}{if TRUE f and g are assumed to return gradients. When FALSE, forward diff is used for approximation.}

\item{nmc}{the number of monte carlo replications}

\item{names}{names of the functions}

\item{seed}{optional. seed for monte carlo draws}

\item{...}{additional arguments passed fd_grad()}
}
\value{
a list with components: C (constantine matrices), principle_grads, contributions, totals, conc, dist
}
\description{
Performs a full concordance analysis between f and g
}
\details{
\code{measure} should be an argument-free function which simulates a draw x ~ p(x) where p is the prior measure. If \code{measure} is numeric, then Monte Carlo draws are simulated from the standard uniform distribution as \code{runif(measure[1])}.
}
