% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_bass.R
\name{conc_bass}
\alias{conc_bass}
\title{Concordance analysis using bass models}
\usage{
conc_bass(
  mod1,
  mod2,
  prior = NULL,
  mcmc.use = NULL,
  type = 1,
  prior_func = NULL,
  func.use = NULL
)
}
\arguments{
\item{mod1}{a fitted BASS model for first function}

\item{mod2}{a fitted BASS model for second function}

\item{prior}{NULL (default) \verb{[0, 1]} prior for each variable. See details for required structure of prior}

\item{mcmc.use}{vector of mcmc indices to be used for both models. Otherwise, a matrix}

\item{type}{Only used if class(mod1) == class(mod2) == "bassBasis". See \code{C_bassPCA}, \code{C_bassPCA_v2}, \code{Cfg_bassPCA}, and \code{Cfg_bassPCA_v2} for details}

\item{prior_func}{Only used if class(mod1) == class(mod2) == "bassBasis". See \code{C_bassPCA}, \code{C_bassPCA_v2}, \code{Cfg_bassPCA}, and \code{Cfg_bassPCA_v2} for details}

\item{func.use}{Only used if class(mod1) == class(mod2) == "bassBasis". See \code{C_bassPCA}, \code{C_bassPCA_v2}, \code{Cfg_bassPCA}, and \code{Cfg_bassPCA_v2} for details}
}
\value{
A list with matrices Cf, Cg, Cfg, Vfg and the concordance.
}
\description{
Closed form estimator of the Cfg matrix using a BASS model
}
\details{
When models are class bass, each field of the returned object is a list for each mcmc iteration (or a vector for concordance). If mcmc.use = NULL or length(mcmc.use) = 1, then each field is just a matrix (or a scalar for concordance).

When models are class bassBasis (from bassPCA function), each field will be a list for each time point in func.use (func.use = NULL uses all time points in the training data by default). Each component of the list has the same structure as described above for the class == "bass" case.
}
