\name{concord-package}
\alias{concord-package}
\alias{concord}
\docType{package}
\title{Product Concordance}
\description{Convert to and from HS, ISIC Rev. 2/3, and SITC1,2,3,4}
\details{
\tabular{ll}{
Package: \tab concord\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-12-26\cr
License: \tab GPL-2 \cr
}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
Feng Zhu <zhufeng@umich.edu>
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ concord }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
