% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concord_bec_uspc.R
\name{concord_bec_uspc}
\alias{concord_bec_uspc}
\title{Converting USPC and BEC Codes}
\source{
Concordance tables provided by:
\itemize{
  \item United States Patent and Trademark Office <https://www.uspto.gov/web/patents/classification/selectnumwithtitle.htm>
  \item United Nations Trade Statistics <https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp>
}
}
\usage{
concord_bec_uspc(sourcevar, origin, destination, dest.digit = 3, all = FALSE)
}
\arguments{
\item{sourcevar}{An input character vector of USPC or BEC codes. The function accepts 3-digit codes for USPC (class), 5 to 11-digit codes for USPC (subclass), and 1 to 3-digit codes for BEC.}

\item{origin}{A string setting the input industry/patent classification: "USPC" (combined), "USPC2012", "BEC4" (2016).}

\item{destination}{A string setting the output industry/patent classification: "USPC" (combined), "USPC2012", "BEC4" (2016).}

\item{dest.digit}{An integer indicating the preferred number of digits for output codes. Allows 3-digit codes for USPC (class), 5 to 11-digit codes for USPC (subclass), and 1 to 3-digit codes for BEC codes. The default is 3 digits.}

\item{all}{Either TRUE or FALSE. If TRUE, the function will return (1) all matched outputs for each input, and (2) the share of occurrences for each matched output among all matched outputs. Users can use the shares as weights for more precise concordances. If FALSE, the function will only return the matched output with the largest share of occurrences (the mode match). If the mode consists of multiple matches, the function will return the first matched output.}
}
\value{
The matched output(s) for each element of the input vector. Either a list object when all = TRUE or a character vector when all = FALSE.
}
\description{
Concords the Unites State Patent Classification codes (USPC 2012) to and from Broad Economic Classification codes (BEC Revision 4) via the bridge of NAICS codes.
}
\examples{
# one-to-one match

concord_bec_uspc("002/2.11", origin = "USPC", destination = "BEC4", dest.digit = 1)

concord_bec_uspc("002/2.11", origin = "USPC", destination = "BEC4", dest.digit = 2)

concord_bec_uspc("22", origin = "BEC4", destination = "USPC", dest.digit = 3)

concord_bec_uspc("22", origin = "BEC4", destination = "USPC", dest.digit = 9)

# multiple-to-one match

concord_bec_uspc(c("002/2.11", "072/47"), origin = "USPC", destination = "BEC4", dest.digit = 1)

concord_bec_uspc(c("1", "2"), origin = "BEC4", destination = "USPC", dest.digit = 3)

# one to multiple matches (all = T)

concord_bec_uspc("002/2.11", origin = "USPC", destination = "BEC4", dest.digit = 2, all = TRUE)

concord_bec_uspc("61", origin = "BEC4", destination = "USPC", dest.digit = 3, all = TRUE)
}
