\name{concord_data}
\alias{concord_data}

\docType{data}
\title{Product Concordance Table}

\description{
Data table containing concordance table for HS, ISIC, and SITC classifications.
Pivots on HS (Combined).
}

\usage{data("concord_data")}

\format{
  A data frame with 6281 observations on the following 7 variables.
  \itemize{
    \item{\code{HS}}
    \item{\code{ISIC2}} 
    \item{\code{ISIC3}} 
    \item{\code{SITC1}} 
    \item{\code{SITC2}} 
    \item{\code{SITC3}} 
    \item{\code{SITC4}}
    }
}
%%\details{}
%%\source{}
%%\references{}

\keyword{ datasets }
