% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs0_to_naics.R
\name{concord_hs0_naics}
\alias{concord_hs0_naics}
\title{Converting HS0 to NAICS Codes}
\usage{
concord_hs0_naics(sourcevar, origin, destination, dest.digit = 6,
  all = FALSE)
}
\arguments{
\item{sourcevar}{An input character vector of HS0 codes. Allows 6, 4, or 2 digits.}

\item{origin}{A string setting the input industry classification (HS0).}

\item{destination}{A string setting the output industry classification (NAICS).}

\item{dest.digit}{An integer indicating the preferred number of digits for output codes. Allows 6, 4, or 2 digits. The default is 6 digits.}

\item{all}{Either TRUE or FALSE. If TRUE, the function will return (1) all matched outputs for each input, and (2) the share of occurrences for each matched output among all matched outputs. Users can use the shares as weights for more precise concordances. If FALSE, the function will only return the matched output with the largest share of occurrences (the mode match). If the mode consists of multiple matches, the function will return the first matched output.}
}
\value{
The matched output(s) for each element of the input vector. Either a list object when all = TRUE or a character vector when all = FALSE.
}
\description{
\code{concord_hs0_naics} converts HS0 to NAICS (combined) codes
}
\examples{
# one input: one-to-one match
concord_hs0_naics(sourcevar = "120600", origin = "HS0", destination = "NAICS", all = FALSE)
concord_hs0_naics(sourcevar = "120600", origin = "HS0", destination = "NAICS", all = TRUE)

# two inputs: multiple-to-one match
concord_hs0_naics(sourcevar = c("120600", "120500"), origin = "HS0", destination = "NAICS", all = FALSE)
concord_hs0_naics(sourcevar = c("120600", "120500"), origin = "HS0", destination = "NAICS", all = TRUE)

# two inputs: repeated
concord_hs0_naics(sourcevar = c("120600", "120600"), origin = "HS0", destination = "NAICS", all = FALSE)

# one to multiple matches
concord_hs0_naics(sourcevar = c("120600", "854690"), origin = "HS0", destination = "NAICS", all = TRUE)

# sourcevar has different number of digits, gives an error
concord_hs0_naics(sourcevar = c("120600", "1206"), origin = "HS0", destination = "NAICS", all = FALSE)

# if no match, will return NA
concord_hs0_naics(sourcevar = c("120600", "120601"), origin = "HS0", destination = "NAICS", all = FALSE)

# 4-digit inputs
concord_hs0_naics(sourcevar = c("1206", "8546"), origin = "HS0", destination = "NAICS", all = TRUE)

# 4-digit outputs
concord_hs0_naics(sourcevar = c("120600", "854690"), origin = "HS0", destination = "NAICS", dest.digit = 4, all = TRUE)
}
