% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concord_hs_ipc.R
\name{concord_hs_ipc}
\alias{concord_hs_ipc}
\title{Converting IPC and HS Codes}
\source{
Concordance tables provided by:
\itemize{
  \item United States Patent and Trademark Office <https://www.uspto.gov/web/patents/classification/selectnumwithtitle.htm>
  \item United Nations Trade Statistics <https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp>
}
}
\usage{
concord_hs_ipc(sourcevar, origin, destination, dest.digit = 4, all = FALSE)
}
\arguments{
\item{sourcevar}{An input character vector of IPC or HS codes. The function accepts 1-digit codes for IPC (section), 3-digit codes for IPC (class), 4-digit codes for IPC (subclass), 9 to 13-digit codes for IPC (group and subgroup), and 2, 4, 6-digit codes for HS.}

\item{origin}{A string setting the input industry/patent classification: "IPC" (combined), "IPC2012", "HS" (combined), "HS0" (1988/92), "HS1" (1996), "HS2" (2002), "HS3" (2007), "HS4" (2012), "HS5" (2017).}

\item{destination}{A string setting the output industry/patent classification: "IPC" (combined), "IPC2012", "HS" (combined), "HS0" (1988/92), "HS1" (1996), "HS2" (2002), "HS3" (2007), "HS4" (2012), "HS5" (2017).}

\item{dest.digit}{An integer indicating the preferred number of digits for output codes. Allows 1-digit codes for IPC (section), 3-digit codes for IPC (class), 4-digit codes for IPC (subclass), 9 to 13-digit codes for IPC (group and subgroup), and 2, 4, 6-digit codes for HS. The default is 4 digits.}

\item{all}{Either TRUE or FALSE. If TRUE, the function will return (1) all matched outputs for each input, and (2) the share of occurrences for each matched output among all matched outputs. Users can use the shares as weights for more precise concordances. If FALSE, the function will only return the matched output with the largest share of occurrences (the mode match). If the mode consists of multiple matches, the function will return the first matched output.}
}
\value{
The matched output(s) for each element of the input vector. Either a list object when all = TRUE or a character vector when all = FALSE.
}
\description{
Concords International Patent Classification codes (IPC 2012) to and from Harmonized System codes (HS0, HS1, HS2, HS3, HS4, HS5, HS combined) via the bridge of USPC and NAICS codes.
}
\examples{
# one-to-one match

concord_hs_ipc("A41D 19/00", origin = "IPC", destination = "HS", dest.digit = 4)

concord_hs_ipc("A41D 19/00", origin = "IPC", destination = "HS", dest.digit = 6)

concord_hs_ipc("A41D 19/00", origin = "IPC", destination = "HS3", dest.digit = 4)

concord_hs_ipc("A41D 19/00", origin = "IPC2012", destination = "HS", dest.digit = 4)

concord_hs_ipc("5205", origin = "HS", destination = "IPC", dest.digit = 4)

concord_hs_ipc("5205", origin = "HS", destination = "IPC", dest.digit = 9)

concord_hs_ipc("5205", origin = "HS", destination = "IPC2012", dest.digit = 4)

concord_hs_ipc("5205", origin = "HS3", destination = "IPC", dest.digit = 4)

# multiple-to-one match

concord_hs_ipc(c("A41D 19/00", "A41F 9/00"), origin = "IPC", destination = "HS", dest.digit = 4)

concord_hs_ipc(c("A41D 19/00", "A41F 9/00"), origin = "IPC", destination = "HS", dest.digit = 6)

concord_hs_ipc(c("500400", "590190"), origin = "HS", destination = "IPC", dest.digit = 4)

concord_hs_ipc(c("500400", "590190"), origin = "HS", destination = "IPC", dest.digit = 9)

# one to multiple matches (all = T)

concord_hs_ipc("A41D 19/00", origin = "IPC", destination = "HS", dest.digit = 4, all = TRUE)

concord_hs_ipc("5901", origin = "HS", destination = "IPC", dest.digit = 4, all = TRUE)
}
