% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_sitc.R
\name{concord_hs_sitc}
\alias{concord_hs_sitc}
\title{Converting HS and SITC Codes}
\usage{
concord_hs_sitc(sourcevar, origin, destination, dest.digit = 4,
  all = FALSE)
}
\arguments{
\item{sourcevar}{An input character vector of HS or SITC codes. The function accepts 2, 4, 6, 10-digit codes for HS; 2, 4, 6-digit codes for HS0, HS1, HS2, HS3, HS4 codes; 1 to 5-digit codes for SITC.}

\item{origin}{A string setting the input industry classification: "HS" (HS combined), "HS0", "HS1", "HS2", "HS3", "HS4", "SITC4".}

\item{destination}{A string setting the output industry classification: "HS" (HS combined), "HS0", "HS1", "HS2", "HS3", "HS4", "SITC4".}

\item{dest.digit}{An integer indicating the preferred number of digits for output codes. Allows 2, 4, 6, or 10 digits for HS; 2, 4, or 6 digits for HS0, HS1, HS2, HS3, HS4 codes; 1 to 5 digits for SITC4 codes. The default is 4 digits.}

\item{all}{Either TRUE or FALSE. If TRUE, the function will return (1) all matched outputs for each input, and (2) the share of occurrences for each matched output among all matched outputs. Users can use the shares as weights for more precise concordances. If FALSE, the function will only return the matched output with the largest share of occurrences (the mode match). If the mode consists of multiple matches, the function will return the first matched output.}
}
\value{
The matched output(s) for each element of the input vector. Either a list object when all = TRUE or a character vector when all = FALSE.
}
\description{
\code{concord_hs_sitc} converts HS codes (HS, HS0, HS1, HS2, HS3, HS4) to and from SITC (Revision 4) codes.
}
\examples{
## HS combined to SITC4
# one input: one-to-one match
concord_hs_sitc(sourcevar = "1206000069", origin = "HS", destination = "SITC4", dest.digit = 5, all = FALSE)
concord_hs_sitc(sourcevar = "1206000069", origin = "HS", destination = "SITC4", dest.digit = 5, all = TRUE)

# two inputs: multiple-to-one match
concord_hs_sitc(sourcevar = c("1206000069", "1206000061"), origin = "HS", destination = "SITC4", dest.digit = 5, all = FALSE)
concord_hs_sitc(sourcevar = c("1206000069", "1206000061"), origin = "HS", destination = "SITC4", dest.digit = 5, all = TRUE)

# two inputs: repeated
concord_hs_sitc(sourcevar = c("1206000069", "1206000069"), origin = "HS", destination = "SITC4", dest.digit = 5, all = FALSE)

# one to multiple matches
concord_hs_sitc(sourcevar = c("1206000069", "0305720000"), origin = "HS", destination = "SITC4", dest.digit = 5, all = TRUE)

# sourcevar has different number of digits, gives an error
concord_hs_sitc(sourcevar = c("1206000069", "120600006"), origin = "HS", destination = "SITC4", dest.digit = 5, all = FALSE)

# sourcevar includes NAs, gives an error
concord_hs_sitc(sourcevar = c("1206000069", NA), origin = "HS", destination = "SITC4", dest.digit = 5, all = FALSE)

# if no match, will return NA
concord_hs_sitc(sourcevar = c("1206000069", "1206000062"), origin = "HS", destination = "SITC4", dest.digit = 5, all = FALSE)

# 5-digit inputs not supported and will give an error
concord_hs_sitc(sourcevar = c("12060", "85460"), origin = "HS", destination = "SITC4", dest.digit = 5, all = TRUE)

# 4-digit inputs, 5-digit outputs
concord_hs_sitc(sourcevar = c("1206", "8546"), origin = "HS", destination = "SITC4", dest.digit = 5, all = TRUE)

# 10-digit inputs, 3-digit outputs
concord_hs_sitc(sourcevar = c("1206000069", "1206000061"), origin = "HS", destination = "SITC4", dest.digit = 3, all = TRUE)

## HS0 to SITC4
concord_hs_sitc(sourcevar = c("1206", "8546"), origin = "HS0", destination = "SITC4", dest.digit = 5, all = TRUE)
concord_hs_sitc(sourcevar = c("120600", "854690"), origin = "HS0", destination = "SITC4", dest.digit = 3, all = TRUE)

## HS1 to SITC4
concord_hs_sitc(sourcevar = c("1206", "8546"), origin = "HS1", destination = "SITC4", dest.digit = 5, all = TRUE)
concord_hs_sitc(sourcevar = c("120600", "854690"), origin = "HS1", destination = "SITC4", dest.digit = 3, all = TRUE)

## HS2 to SITC4
concord_hs_sitc(sourcevar = c("1206", "8546"), origin = "HS2", destination = "SITC4", dest.digit = 5, all = TRUE)
concord_hs_sitc(sourcevar = c("120600", "854690"), origin = "HS2", destination = "SITC4", dest.digit = 3, all = TRUE)

## HS3 to SITC4
concord_hs_sitc(sourcevar = c("1206", "8546"), origin = "HS3", destination = "SITC4", dest.digit = 5, all = TRUE)
concord_hs_sitc(sourcevar = c("120600", "854690"), origin = "HS3", destination = "SITC4", dest.digit = 3, all = TRUE)

## HS4 to SITC4
concord_hs_sitc(sourcevar = c("1206", "8546"), origin = "HS4", destination = "SITC4", dest.digit = 5, all = TRUE)
concord_hs_sitc(sourcevar = c("120600", "854690"), origin = "HS4", destination = "SITC4", dest.digit = 3, all = TRUE)

## SITC4 to HS
concord_hs_sitc(sourcevar = c("2224", "7732"), origin = "SITC4", destination = "HS", dest.digit = 6, all = TRUE)
concord_hs_sitc(sourcevar = c("22240", "77324"), origin = "SITC4", destination = "HS", dest.digit = 6, all = TRUE)

## SITC4 to HS0
concord_hs_sitc(sourcevar = c("2224", "7732"), origin = "SITC4", destination = "HS0", dest.digit = 6, all = TRUE)
concord_hs_sitc(sourcevar = c("22240", "77324"), origin = "SITC4", destination = "HS0", dest.digit = 6, all = TRUE)

## SITC4 to HS1
concord_hs_sitc(sourcevar = c("2224", "7732"), origin = "SITC4", destination = "HS1", dest.digit = 6, all = TRUE)
concord_hs_sitc(sourcevar = c("22240", "77324"), origin = "SITC4", destination = "HS1", dest.digit = 6, all = TRUE)

## SITC4 to HS2
concord_hs_sitc(sourcevar = c("2224", "7732"), origin = "SITC4", destination = "HS2", dest.digit = 6, all = TRUE)
concord_hs_sitc(sourcevar = c("22240", "77324"), origin = "SITC4", destination = "HS2", dest.digit = 6, all = TRUE)

## SITC4 to HS3
concord_hs_sitc(sourcevar = c("2224", "7732"), origin = "SITC4", destination = "HS3", dest.digit = 6, all = TRUE)
concord_hs_sitc(sourcevar = c("22240", "77324"), origin = "SITC4", destination = "HS3", dest.digit = 6, all = TRUE)

## SITC4 to HS4
concord_hs_sitc(sourcevar = c("2224", "7732"), origin = "SITC4", destination = "HS4", dest.digit = 6, all = TRUE)
concord_hs_sitc(sourcevar = c("22240", "77324"), origin = "SITC4", destination = "HS4", dest.digit = 6, all = TRUE)
}
