% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concord_isic_ipc.R
\name{concord_isic_ipc}
\alias{concord_isic_ipc}
\title{Converting IPC and ISIC Codes}
\source{
Concordance tables provided by:
\itemize{ 
  \item United States Patent and Trademark Office <https://www.uspto.gov/web/patents/classification/selectnumwithtitle.htm>
  \item United Nations Trade Statistics <https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp>
}
}
\usage{
concord_isic_ipc(sourcevar, origin, destination, dest.digit = 4, all = FALSE)
}
\arguments{
\item{sourcevar}{An input character vector of IPC or ISIC codes. The function accepts 1-digit codes for IPC (section), 3-digit codes for IPC (class), 4-digit codes for IPC (subclass), 9 to 13-digit codes for IPC (group and subgroup), and 1 to 4-digit codes for ISIC.}

\item{origin}{A string setting the input industry/patent classification: "IPC" (combined), "IPC2012", "ISIC2" (1968), "ISIC3" (1989), "ISIC3.1" (2002), "ISIC4" (2008).}

\item{destination}{A string setting the output industry/patent classification: "IPC" (combined), "IPC2012", "ISIC2" (1968), "ISIC3" (1989), "ISIC3.1" (2002), "ISIC4" (2008).}

\item{dest.digit}{An integer indicating the preferred number of digits for output codes. Allows 1-digit codes for IPC (section), 3-digit codes for IPC (class), 4-digit codes for IPC (subclass), 9 to 13-digit codes for IPC (group and subgroup), and 1 to 4-digit codes for ISIC codes. The default is 4 digits.}

\item{all}{Either TRUE or FALSE. If TRUE, the function will return (1) all matched outputs for each input, and (2) the share of occurrences for each matched output among all matched outputs. Users can use the shares as weights for more precise concordances. If FALSE, the function will only return the matched output with the largest share of occurrences (the mode match). If the mode consists of multiple matches, the function will return the first matched output.}
}
\value{
The matched output(s) for each element of the input vector. Either a list object when all = TRUE or a character vector when all = FALSE.
}
\description{
Concords International Patent Classification codes (IPC 2012) to and from International Standard Industrial Classification codes (ISIC Revision 2, 3, 3.1, 4) via the bridge of USPC and NAICS codes.
}
\examples{
# one-to-one match

concord_isic_ipc("A41D 19/00", origin = "IPC", destination = "ISIC3", dest.digit = 3)

concord_isic_ipc("A41D 19/00", origin = "IPC", destination = "ISIC3", dest.digit = 4)

concord_isic_ipc("A41D 19/00", origin = "IPC", destination = "ISIC4", dest.digit = 3)

concord_isic_ipc("A41D 19/00", origin = "IPC2012", destination = "ISIC3", dest.digit = 3)

concord_isic_ipc("1711", origin = "ISIC3", destination = "IPC", dest.digit = 4)

concord_isic_ipc("1711", origin = "ISIC3", destination = "IPC", dest.digit = 9)

concord_isic_ipc("1711", origin = "ISIC3", destination = "IPC2012", dest.digit = 4)

concord_isic_ipc("1711", origin = "ISIC3.1", destination = "IPC", dest.digit = 4)

# multiple-to-one match

concord_isic_ipc(c("A41D", "A41F"), origin = "IPC", destination = "ISIC3", dest.digit = 3)

concord_isic_ipc(c("A41D", "A41F"), origin = "IPC", destination = "ISIC3", dest.digit = 4)

concord_isic_ipc(c("1711", "1721"), origin = "ISIC3.1", destination = "IPC", dest.digit = 4)

concord_isic_ipc(c("1711", "1721"), origin = "ISIC3.1", destination = "IPC", dest.digit = 9)

# one to multiple matches (all = T)

concord_isic_ipc("A41D 19/00", origin = "IPC", destination = "ISIC4", dest.digit = 4, all = TRUE)

concord_isic_ipc("2826", origin = "ISIC4", destination = "IPC", dest.digit = 4, all = TRUE)
}
