% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naics2017_to_hs.R
\name{concord_naics2017_hs}
\alias{concord_naics2017_hs}
\title{Converts NAICS 2017 to HS Codes}
\usage{
concord_naics2017_hs(sourcevar, origin, destination, dest.digit = 6,
  all = FALSE)
}
\arguments{
\item{sourcevar}{An input character vector of NAICS 2017 codes. Allows 6, 4, or 2 digits.}

\item{origin}{A string setting the input industry classification (NAICS2017).}

\item{destination}{A string setting the output industry classification (HS).}

\item{dest.digit}{An integer indicating the preferred number of digits for output codes. Allows 10, 6, 4, or 2 digits. The default is 6 digits.}

\item{all}{Either TRUE or FALSE. If TRUE, the function will return (1) all matched outputs for each input, and (2) the share of occurrences for each matched output among all matched outputs. Users can use the shares as weights for more precise concordances. If FALSE, the function will only return the matched output with the largest share of occurrences (the mode match). If the mode consists of multiple matches, the function will return the first matched output.}
}
\value{
The matched output(s) for each element of the input vector. Either a list object when all = TRUE or a character vector when all = FALSE.
}
\description{
\code{concord_naics2017_hs} converts NAICS 2017 to HS codes
}
\examples{
# one input: one-to-multiple match
concord_naics2017_hs(sourcevar = "111120", origin = "NAICS2017", destination = "HS", all = FALSE)
concord_naics2017_hs(sourcevar = "111120", origin = "NAICS2017", destination = "HS", all = TRUE)

# two inputs: one-to-multiple match
concord_naics2017_hs(sourcevar = c("111120", "326199"), origin = "NAICS2017", destination = "HS", all = FALSE)
concord_naics2017_hs(sourcevar = c("111120", "326199"), origin = "NAICS2017", destination = "HS", all = TRUE)

# two inputs: repeated
concord_naics2017_hs(sourcevar = c("111120", "111120"), origin = "NAICS2017", destination = "HS", all = FALSE)
concord_naics2017_hs(sourcevar = c("111120", "111120"), origin = "NAICS2017", destination = "HS", all = TRUE)

# sourcevar has different number of digits, will give an error
concord_naics2017_hs(sourcevar = c("111120", "11112"), origin = "NAICS2017", destination = "HS", all = FALSE)

# If no match, will return NA
concord_naics2017_hs(sourcevar = c("111120", "111121"), origin = "NAICS2017", destination = "HS", all = FALSE)

# 4-digit inputs
concord_naics2017_hs(sourcevar = c("1111", "3271"), origin = "NAICS2017", destination = "HS", all = FALSE)

# 4-digit outputs
concord_naics2017_hs(sourcevar = c("111120", "326199"), origin = "NAICS2017", destination = "HS", dest.digit = 4, all = FALSE)
}
