% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concord_naics_ipc.R
\name{concord_naics_ipc}
\alias{concord_naics_ipc}
\title{Converting IPC and NAICS Codes}
\source{
Concordance tables provided by:
\itemize{
  \item United States Patent and Trademark Office <https://www.uspto.gov/web/patents/classification/selectnumwithtitle.htm>
}
}
\usage{
concord_naics_ipc(sourcevar, origin, destination, dest.digit = 4, all = FALSE)
}
\arguments{
\item{sourcevar}{An input character vector of IPC or NAICS codes. The function accepts 1-digit codes for IPC (section), 3-digit codes for IPC (class), 4-digit codes for IPC (subclass), 9 to 13-digit codes for IPC (group and subgroup), and 2 to 6-digit codes for NAICS.}

\item{origin}{A string setting the input industry/patent classification: "IPC" (combined), "IPC2012", "NAICS2002", "NAICS2007", "NAICS2012", "NAICS2017", "NAICS" (combined).}

\item{destination}{A string setting the output industry/patent classification: "IPC" (combined), "IPC2012", "NAICS2002", "NAICS2007", "NAICS2012", "NAICS2017", "NAICS" (combined).}

\item{dest.digit}{An integer indicating the preferred number of digits for output codes. Allows 1-digit codes for IPC (section), 3-digit codes for IPC (class), 4-digit codes for IPC (subclass), 9 to 13-digit codes for IPC (group and subgroup), and 2 to 6-digit codes for NAICS codes. The default is 4 digits.}

\item{all}{Either TRUE or FALSE. If TRUE, the function will return (1) all matched outputs for each input, and (2) the share of occurrences for each matched output among all matched outputs. Users can use the shares as weights for more precise concordances. If FALSE, the function will only return the matched output with the largest share of occurrences (the mode match). If the mode consists of multiple matches, the function will return the first matched output.}
}
\value{
The matched output(s) for each element of the input vector. Either a list object when all = TRUE or a character vector when all = FALSE.
}
\description{
Concords International Patent Classification codes (IPC 2012) to and from North American Industry Classification System codes (NAICS2002, NAICS2007, NAICS2012, NAICS2017, NAICS combined) via the bridge of USPC codes.
}
\examples{
# one-to-one match

concord_naics_ipc("A41D 19/00", origin = "IPC", destination = "NAICS", dest.digit = 3)

concord_naics_ipc("A41D 19/00", origin = "IPC", destination = "NAICS", dest.digit = 4)

concord_naics_ipc("A41D 19/00", origin = "IPC", destination = "NAICS2002", dest.digit = 3)

concord_naics_ipc("A41D 19/00", origin = "IPC2012", destination = "NAICS", dest.digit = 3)

concord_naics_ipc("3133", origin = "NAICS", destination = "IPC", dest.digit = 4)

concord_naics_ipc("3133", origin = "NAICS", destination = "IPC", dest.digit = 9)

concord_naics_ipc("3133", origin = "NAICS", destination = "IPC2012", dest.digit = 4)

concord_naics_ipc("3133", origin = "NAICS2007", destination = "IPC", dest.digit = 4)

# multiple-to-one match

concord_naics_ipc(c("A41D", "A41F"), origin = "IPC", destination = "NAICS", dest.digit = 3)

concord_naics_ipc(c("A41D", "A41F"), origin = "IPC", destination = "NAICS", dest.digit = 4)

concord_naics_ipc(c("313", "333"), origin = "NAICS", destination = "IPC", dest.digit = 4)

concord_naics_ipc(c("313", "333"), origin = "NAICS", destination = "IPC", dest.digit = 9)

# one to multiple matches (all = T)

concord_naics_ipc("A41D 19/00", origin = "IPC", destination = "NAICS", dest.digit = 3, all = TRUE)

concord_naics_ipc("3131", origin = "NAICS", destination = "IPC", dest.digit = 4, all = TRUE)
}
