% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concord_naics_isic.R
\name{concord_naics_isic}
\alias{concord_naics_isic}
\title{Converting NAICS and ISIC Codes}
\source{
Concordance tables provided by:
\itemize{
  \item The U.S. Census Bureau <https://www.census.gov/eos/www/naics/concordances/concordances.html>
}
}
\usage{
concord_naics_isic(sourcevar, origin, destination, dest.digit = 4, all = FALSE)
}
\arguments{
\item{sourcevar}{An input character vector of NAICS or ISIC codes. The function accepts 2, 4, 6-digit codes for NAICS and 1 to 4-digit codes for ISIC.}

\item{origin}{A string setting the input industry classification: "NAICS2002", "NAICS2007", "NAICS2012", "NAICS2017", "ISIC2" (1968), "ISIC3" (1989), "ISIC3.1" (2002), "ISIC4" (2008).}

\item{destination}{A string setting the output industry classification: "NAICS2002", "NAICS2007", "NAICS2012", "NAICS2017", "ISIC2" (1968), "ISIC3" (1989), "ISIC3.1" (2002), "ISIC4" (2008).}

\item{dest.digit}{An integer indicating the preferred number of digits for output codes. Allows 2, 4, or 6 digits for NAICS codes and 1 to 4 digits for ISIC codes. The default is 4 digits.}

\item{all}{Either TRUE or FALSE. If TRUE, the function will return (1) all matched outputs for each input, and (2) the share of occurrences for each matched output among all matched outputs. Users can use the shares as weights for more precise concordances. If FALSE, the function will only return the matched output with the largest share of occurrences (the mode match). If the mode consists of multiple matches, the function will return the first matched output.}
}
\value{
The matched output(s) for each element of the input vector. Either a list object when all = TRUE or a character vector when all = FALSE.
}
\description{
Concords North American Industry Classification System codes (NAICS2002, NAICS2007, NAICS2012, NAICS2017) to and from International Standard Industrial Classification codes (ISIC Revision 2, 3, 3.1, 4).
}
\note{
Always include leading zeros in codes (e.g., use ISIC4 code 0111 instead of 111)---results may be buggy otherwise.
}
\examples{
## NAICS2017 to ISIC4
concord_naics_isic(sourcevar = c("111120", "326199"),
                origin = "NAICS2017", destination = "ISIC4",
                dest.digit = 4, all = TRUE)
}
