% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitc4_to_hs.R
\name{concord_sitc4_hs}
\alias{concord_sitc4_hs}
\title{Converts SITC4 to HS Codes}
\usage{
concord_sitc4_hs(sourcevar, origin, destination, dest.digit = 6,
  all = FALSE)
}
\arguments{
\item{sourcevar}{An input character vector of SITC4 codes. Allows 5, 4, 3, 2, or 1 digits.}

\item{origin}{A string setting the input industry classification (SITC4).}

\item{destination}{A string setting the output industry classification (HS).}

\item{dest.digit}{An integer indicating the preferred number of digits for output codes. Allows 10, 6, 4, or 2 digits. The default is 6 digits.}

\item{all}{Either TRUE or FALSE. If TRUE, the function will return (1) all matched outputs for each input, and (2) the share of occurrences for each matched output among all matched outputs. Users can use the shares as weights for more precise concordances. If FALSE, the function will only return the matched output with the largest share of occurrences (the mode match). If the mode consists of multiple matches, the function will return the first matched output.}
}
\value{
The matched output(s) for each element of the input vector. Either a list object when all = TRUE or a character vector when all = FALSE.
}
\description{
\code{concord_sitc4_hs} converts SITC4 Revision 4 to HS (combined) codes.
}
\examples{
# one input: one-to-multiple match
concord_sitc4_hs(sourcevar = "22240", origin = "SITC4", destination = "HS", all = FALSE)
concord_sitc4_hs(sourcevar = "22240", origin = "SITC4", destination = "HS", all = TRUE)

# two inputs: one-to-multiple match
concord_sitc4_hs(sourcevar = c("22240", "00119"), origin = "SITC4", destination = "HS", all = FALSE)
concord_sitc4_hs(sourcevar = c("22240", "00119"), origin = "SITC4", destination = "HS", all = TRUE)

# two inputs: repeated
concord_sitc4_hs(sourcevar = c("22240", "22240"), origin = "SITC4", destination = "HS", all = FALSE)
concord_sitc4_hs(sourcevar = c("22240", "22240"), origin = "SITC4", destination = "HS", all = TRUE)

# sourcevar has different number of digits, will give an error
concord_sitc4_hs(sourcevar = c("22240", "2224"), origin = "SITC4", destination = "HS", all = FALSE)

# If no match, will return NA
concord_sitc4_hs(sourcevar = c("22240", "22241"), origin = "SITC4", destination = "HS", all = FALSE)

# 4-digit inputs
concord_sitc4_hs(sourcevar = c("2224", "0011"), origin = "SITC4", destination = "HS", all = FALSE)

# 4-digit outputs
concord_sitc4_hs(sourcevar = c("22240", "00119"), origin = "SITC4", destination = "HS", dest.digit = 4, all = FALSE)
}
