% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concord_sitc_ipc.R
\name{concord_sitc_ipc}
\alias{concord_sitc_ipc}
\title{Converting IPC and SITC Codes}
\source{
Concordance tables provided by:
\itemize{
  \item United States Patent and Trademark Office <https://www.uspto.gov/web/patents/classification/selectnumwithtitle.htm>
  \item United Nations Trade Statistics <https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp>
}
}
\usage{
concord_sitc_ipc(sourcevar, origin, destination, dest.digit = 4, all = FALSE)
}
\arguments{
\item{sourcevar}{An input character vector of IPC or SITC codes. The function accepts 1-digit codes for IPC (section), 3-digit codes for IPC (class), 4-digit codes for IPC (subclass), 9 to 13-digit codes for IPC (group and subgroup), and 1 to 5-digit codes for SITC.}

\item{origin}{A string setting the input industry/patent classification: "IPC" (combined), "IPC2012", "SITC1" (1950), "SITC2" (1974), "SITC3" (1985), "SITC4" (2006).}

\item{destination}{A string setting the output industry/patent classification: "IPC" (combined), "IPC2012", "SITC1" (1950), "SITC2" (1974), "SITC3" (1985), "SITC4" (2006).}

\item{dest.digit}{An integer indicating the preferred number of digits for output codes. Allows 1-digit codes for IPC (section), 3-digit codes for IPC (class), 4-digit codes for IPC (subclass), 9 to 13-digit codes for IPC (group and subgroup), and 1 to 5-digit codes for SITC. The default is 4 digits.}

\item{all}{Either TRUE or FALSE. If TRUE, the function will return (1) all matched outputs for each input, and (2) the share of occurrences for each matched output among all matched outputs. Users can use the shares as weights for more precise concordances. If FALSE, the function will only return the matched output with the largest share of occurrences (the mode match). If the mode consists of multiple matches, the function will return the first matched output.}
}
\value{
The matched output(s) for each element of the input vector. Either a list object when all = TRUE or a character vector when all = FALSE.
}
\description{
Concords International Patent Classification codes (IPC 2012) to and from Standard International Trade Classification codes (SITC1, SITC2, SITC3, SITC4) via the bridge of USPC and NAICS codes.
}
\examples{
# one-to-one match
concord_sitc_ipc("A41D 19/00", origin = "IPC", destination = "SITC3", dest.digit = 3)

concord_sitc_ipc("A41D 19/00", origin = "IPC", destination = "SITC3", dest.digit = 5)

concord_sitc_ipc("A41D 19/00", origin = "IPC", destination = "SITC4", dest.digit = 3)

concord_sitc_ipc("A41D 19/00", origin = "IPC2012", destination = "SITC3", dest.digit = 3)

concord_sitc_ipc("65112", origin = "SITC3", destination = "IPC", dest.digit = 4)

concord_sitc_ipc("65112", origin = "SITC3", destination = "IPC", dest.digit = 9)

concord_sitc_ipc("65112", origin = "SITC3", destination = "IPC2012", dest.digit = 4)

concord_sitc_ipc("65112", origin = "SITC4", destination = "IPC", dest.digit = 4)

# multiple-to-one match

concord_sitc_ipc(c("A41D", "A41F"), origin = "IPC", destination = "SITC3", dest.digit = 3)

concord_sitc_ipc(c("A41D", "A41F"), origin = "IPC", destination = "SITC3", dest.digit = 4)

concord_sitc_ipc(c("65112", "65113"), origin = "SITC3", destination = "IPC", dest.digit = 4)

concord_sitc_ipc(c("65112", "65113"), origin = "SITC3", destination = "IPC", dest.digit = 9)

# one to multiple matches (all = T)

concord_sitc_ipc("A41D 19/00", origin = "IPC", destination = "SITC3", dest.digit = 4, all = TRUE)

concord_sitc_ipc("65112", origin = "SITC4", destination = "IPC", dest.digit = 4, all = TRUE)
}
