% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concord_sitc_isic.R
\name{concord_sitc_isic}
\alias{concord_sitc_isic}
\title{Converting SITC and ISIC Codes}
\source{
Concordance tables provided by:
\itemize{
  \item World Integrated Trade Solution (WITS), World Bank <https://wits.worldbank.org/product_concordance.html>
}
}
\usage{
concord_sitc_isic(sourcevar, origin, destination, dest.digit = 4, all = FALSE)
}
\arguments{
\item{sourcevar}{An input character vector of SITC or ISIC codes. The function accepts 1 to 5-digit codes for SITC and 1 to 4-digit codes for ISIC.}

\item{origin}{A string setting the input industry classification: "SITC1" (1950), "SITC2" (1974), "SITC3" (1985), "SITC4" (2006), "ISIC2" (1968), "ISIC3" (1989), "ISIC3.1" (2002), "ISIC4" (2008).}

\item{destination}{A string setting the output industry classification: "SITC1" (1950), "SITC2" (1974), "SITC3" (1985), "SITC4" (2006), "ISIC2" (1968), "ISIC3" (1989), "ISIC3.1" (2002), "ISIC4" (2008).}

\item{dest.digit}{An integer indicating the preferred number of digits for output codes. Allows 1 to 5 digits for SITC codes and 1 to 4 digits for ISIC codes. The default is 4 digits.}

\item{all}{Either TRUE or FALSE. If TRUE, the function will return (1) all matched outputs for each input, and (2) the share of occurrences for each matched output among all matched outputs. Users can use the shares as weights for more precise concordances. If FALSE, the function will only return the matched output with the largest share of occurrences (the mode match). If the mode consists of multiple matches, the function will return the first matched output.}
}
\value{
The matched output(s) for each element of the input vector. Either a list object when all = TRUE or a character vector when all = FALSE.
}
\description{
Concords Standard International Trade Classification codes (SITC Revision 1, 2, 3, 4) to and from International Standard Industrial Classification codes (ISIC Revision 2, 3, 3.1, 4).
}
\note{
Always include leading zeros in codes (e.g., use HS code 010110 instead of 10110)---results may be buggy otherwise.
}
\examples{
## SITC3 to ISIC3
concord_sitc_isic(sourcevar = c("22240", "04110"),
                  origin = "SITC3", destination = "ISIC3",
                  dest.digit = 4, all = TRUE)
}
