\name{concord_test}
\alias{concord_test}

\title{concord_test}
\description{Test function for product concordance package}
\usage{concord_test()}

%% \details{}

\value{ Returns TRUE if working properly, or FALSE otherwise }

\author{Feng Zhu}
\note{Does not currently perform any tests}

\examples{
## concord_test()
}

\keyword{ test }