% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concord_v2.R
\name{concord_v2}
\alias{concord_v2}
\title{Concording Different Classification Codes}
\usage{
concord_v2(sourcevar, origin, destination, dest.digit = 6, all = FALSE)
}
\arguments{
\item{sourcevar}{An input character vector of industry/product codes to be converted.}

\item{origin}{A string setting the input coding scheme. Currently supports: HS, NAICS2017.}

\item{destination}{A string setting the output coding scheme. Currently supports: HS, NAICS2017.}

\item{dest.digit}{An integer indicating the preferred number of digits for outputs. The default is 6 digits.}

\item{all}{Either TRUE or FALSE. If TRUE, the function will return (1) all matched outputs for each input, and (2) the share of occurrences for each matched output among all matched outputs. Users can use the shares as weights for more precise concordances. If FALSE, the function will only return the matched output with the largest share of occurrences (the mode match). If the mode consists of multiple matches, the function will return the first matched output.}
}
\value{
The matched output(s) for each element of the input vector. Either a list object when all = TRUE or a character vector when all = FALSE.
}
\description{
\code{concord_v2} concords different classification codes used in international trade.
}
\examples{
# HS code to NAICS 2017 code
concord_v2(sourcevar = c("1206000069", "8546900000"), origin = "HS", destination = "NAICS2017", dest.digit = 6, all = FALSE)
concord_v2(sourcevar = c("1206000069", "8546900000"), origin = "HS", destination = "NAICS2017", dest.digit = 6, all = TRUE)
concord_v2(sourcevar = c("1206000069", "8546900000"), origin = "HS", destination = "NAICS2017", dest.digit = 4, all = TRUE)

# NAICS 2017 code to HS code
concord_v2(sourcevar = c("111120", "326199"), origin = "NAICS2017", destination = "HS", dest.digit = 10, all = FALSE)
concord_v2(sourcevar = c("111120", "326199"), origin = "NAICS2017", destination = "HS", dest.digit = 10, all = TRUE)
concord_v2(sourcevar = c("111120", "326199"), origin = "NAICS2017", destination = "HS", dest.digit = 6, all = FALSE)
}
