\name{concordance-package}
\alias{concord-package}

\docType{package}
\title{Product Concordance}
\description{Convert between HS0/1/2/3/4/Combined, ISIC Rev. 2/3/3.1/4, SITC1,2,3,4, NAICS, SIC, and BEC; also code nomenclature and trade elasticity look-up.}
\details{
\tabular{ll}{
Package: \tab concordance\cr
Type: \tab Package\cr
Version: \tab 1.7\cr
Date: \tab 2017-02-12\cr
License: \tab GPL-2 \cr
}}

\author{Feng Zhu <zhufeng@umich.edu>}

\references{
Data consolidated from
- World Integrated Trade Solution (WITS), World Bank (http://wits.worldbank.org/product_concordance.html)
- UN Statistics Division (http://unstats.un.org/unsd/trade/conversions/HS%20Correlation%20and%20Conversion%20tables.htm)
- Schott and Pierce's site (http://faculty.som.yale.edu/peterschott/sub_international.htm)
- Broda and Weinstein's site (http://www.columbia.edu/~dew35/TradeElasticities/TradeElasticities.html)

See also
- Justin R. Pierce and Peter K. Schott, "Concording U.S. Harmonized System Codes Over Time".
- C. Broda and D. Weinstein, "Globalization and the Gains from Variety,"
 Quarterly Journal of Economics Volume 121, Issue 2 - May 2006
}

% \examples{}

\keyword{ concord, concordance, rauch, sigma, nomenclature, trade }
