% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/concordance.r
\name{concordance}
\alias{concordance}
\title{Calculate the concordance between two data sets.}
\usage{
concordance(x1, x2, eigen_basis = FALSE)
}
\arguments{
\item{x1}{the data set to compare to a reference (normalizing) data set}

\item{x2}{the reference data set}

\item{eigen_basis}{should \code{x1} and \code{x2} be projected onto
\code{x2}'s basis? Default is \code{FALSE}.}
}
\value{
the concordance value between the two data sets
}
\description{
The \code{concordance} function takes two data sets (either matrices
or data.frames) and calculates the concordance between the two.
}
\examples{
concordance(iris[iris$Species=="setosa",c("Sepal.Length", "Sepal.Width")],
            iris[iris$Species=="virginica",c("Sepal.Length","Sepal.Width")]
}

