% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/concordance_test.r
\name{concordance_diff_test}
\alias{concordance_diff_test}
\title{Test the difference of two data sets based on their concordance}
\usage{
concordance_diff_test(x1, x2, alternative = "two.sided", model = "all",
  conf.level = 0.95, eigen_basis = FALSE)
}
\arguments{
\item{x1}{the data set to compare to a reference (normalizing) data set}

\item{x2}{the reference data set}

\item{alternative}{"two.sided", "less", or "greater. Default is "two.sided".}

\item{model}{"Beta", "F", "Cauchy", or "all". The default is "all".
A non-parametric test is coming.}

\item{conf.level}{confidence level for the interval}

\item{eigen_basis}{should \code{x1} and \code{x2} be projected onto
\code{x2}'s basis? Default is \code{FALSE}.}
}
\description{
Test the difference of two data sets based on their concordance
}

