\name{desclen}
\alias{desclen}

\docType{data}
\title{Product code lengths}

\description{
Named list containing full length of various product classification schemes
}

%%  \usage{data("lengths")}

\format{
  A 14 by 6 matrix, with columns indexed by lengths (1 through 6) and rows named
  \itemize{
    \item{\code{HS}}
    \item{\code{HS0}}
    \item{\code{HS1}}
    \item{\code{HS2}}
    \item{\code{HS3}}
    \item{\code{HS4}}
    \item{\code{ISIC2}} 
    \item{\code{ISIC3}} 
    \item{\code{SITC1}} 
    \item{\code{SITC2}} 
    \item{\code{SITC3}} 
    \item{\code{SITC4}}
    \item{\code{BEC}} 
    \item{\code{NAICS}}
    }
}

%%\details{}

\source{
Data consolidated from
- World Integrated Trade Solution (WITS), World Bank (http://wits.worldbank.org/product_concordance.html)
- UN Statistics Division (http://unstats.un.org/unsd/trade/conversions/HS%20Correlation%20and%20Conversion%20tables.htm)
- Schott and Pierce's site (http://faculty.som.yale.edu/peterschott/sub_international.htm)
}

\references{See also Justin R. Pierce and Peter K. Schott, "Concording U.S. Harmonized System Codes Over Time".}

\keyword{ datasets }
