% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{f_piston}
\alias{f_piston}
\title{Piston Function}
\usage{
f_piston(x)
}
\arguments{
\item{x}{7 inputs. See Constantine paper for details}
}
\value{
Time to fire for piston
}
\description{
Piston function studied by Constantine in global sensitivity metrics paper
}
\details{
PARAMETER RANGES:
measure <- function(){
res <- c(
runif(1, 30, 60),
runif(1, .005, .02),
runif(1, .002, .01),
runif(1, 1000, 5000),
runif(1, 90000, 110000),
runif(1, 290, 296),
runif(1, 340, 360)
)
return(res)
}
}
