% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance_r.R
\name{fd_grad}
\alias{fd_grad}
\title{Forward diff function}
\usage{
fd_grad(f, x, h = 1e-12, ...)
}
\arguments{
\item{f}{the function to find the gradient of}

\item{x}{the input values}

\item{h}{the tolerance}

\item{...}{additional inputs to be passed to f}
}
\value{
The approximate gradient of f at x
}
\description{
Function for approximating the gradient of a function
}
