% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_desc.R
\name{get_desc}
\alias{get_desc}
\title{Looking Up Product Description}
\usage{
get_desc(sourcevar, origin)
}
\arguments{
\item{sourcevar}{A character vector of input codes.}

\item{origin}{A string indicating one of the following industry/product classifications: HS, HS0, HS1, HS2, HS3, HS4, ISIC2, ISIC3, SITC1, SITC2, SITC3, SITC4, BEC, NAICS2012, NAICS2017. Note that descriptions for 10-digit HS codes are from the U.S. Census Bureau and thus are only applicable for U.S. trade data.}
}
\value{
A character vector giving the title/description of each element of the input codes.
}
\description{
\code{get_desc} returns the description of product codes
}
\examples{
# NAICS 2017
get_desc(sourcevar = c("111120", "326199", "111120"), origin = "naics2017")

# Returns NA when there are no matches and gives warning
get_desc(sourcevar = c("111121", "326199", "111120", "111120"), origin = "naics2017")

# NAICS 2012
get_desc(sourcevar = c("111120", "326199", "111120"), origin = "naics2012")

# HS
get_desc(sourcevar = c("120600", "854690"), origin = "HS")

# HS0
get_desc(sourcevar = c("120600", "854690"), origin = "HS0")

# HS1
get_desc(sourcevar = c("120600", "854690"), origin = "HS1")

# HS2
get_desc(sourcevar = c("120600", "854690"), origin = "HS2")

# HS3
get_desc(sourcevar = c("120600", "854690"), origin = "HS3")

# HS4
get_desc(sourcevar = c("120600", "854690"), origin = "HS4")

# ISIC2
get_desc(sourcevar = c("3114", "3831"), origin = "ISIC2")

# ISIC3
get_desc(sourcevar = c("1512", "3110"), origin = "ISIC3")

# SITC1
get_desc(sourcevar = c("4216", "7232"), origin = "SITC1")

# SITC2
get_desc(sourcevar = c("4236", "7732"), origin = "SITC2")

# SITC3
get_desc(sourcevar = c("4221", "7732"), origin = "SITC3")

# SITC4
get_desc(sourcevar = c("4221", "7732"), origin = "SITC4")

# BEC
get_desc(sourcevar = c("001", "111"), origin = "BEC")
}
