% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intermediate.R
\name{get_intermediate}
\alias{get_intermediate}
\title{Looking Up the Level of Intermediate Goods Production}
\source{
Product descriptions consolidated from
\itemize{
  \item The U.S. Census Bureau <https://www.census.gov/>
  \item The U.S. Bureau of Labor Statistics <https://www.bls.gov/>
  \item UN Comtrade <https://comtrade.un.org/>
  \item UN Trade Statistics <https://unstats.un.org/unsd/trade/default.asp>
}
}
\usage{
get_intermediate(sourcevar, origin)
}
\arguments{
\item{sourcevar}{An input character vector of industry codes to look up.}

\item{origin}{A string indicating one of the following industry/product classifications: "HS0" (1988/92), "HS1" (1996), "HS2" (2002), "HS3" (2007), "HS4" (2012), "HS5" (2017), "HS6" (2022), "HS" (combined), "SITC1" (1950), "SITC2" (1974), "SITC3" (1985), "SITC4" (2006), "NAICS2002", "NAICS2007", "NAICS2012", "NAICS2017", "ISIC2" (1968), "ISIC3" (1989), "ISIC4" (2008), "BEC4" (2016).}
}
\value{
Uses keywords ("part(s)", "intermediate", and "component") to identify intermediate-goods producing industries (at the most disaggregated level in the description data), and then calculates and returns the proportion these industries occupy among each input code.
}
\description{
Calculates and returns the level (proportion) of intermediate goods production in an industry based on product descriptions.
}
\note{
Please include leading zeros in codes (e.g., use HS code 010110 instead of 10110). For BEC4 only, use original codes or add trailing zeroes if necessary (e.g., 7 or 700 instead of 007). Also note that the results may not be informative for broad categories like BEC4.
}
\examples{
# NAICS
get_intermediate(sourcevar = c("11", "31-33", "42"), origin = "NAICS2017")
get_intermediate(sourcevar = c("3131", "3363"), origin = "NAICS2017")

# HS
get_intermediate(sourcevar = c("03", "84"), origin = "HS5")

# SITC
get_intermediate(sourcevar = c("05", "75"), origin = "SITC4")
}
