% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lifecycle.R
\name{get_lifecycle}
\alias{get_lifecycle}
\title{Looking Up Product Lifecycle}
\source{
Data from NBER Patent Data Project and USPTO report:
\itemize{
  \item{<https://sites.google.com/site/patentdataproject/>}
  \item{<https://www.uspto.gov/web/offices/ac/ido/oeip/taf/naics/doc/naics_info.htm>}
}
}
\usage{
get_lifecycle(sourcevar, origin, setting = "MeanACL")
}
\arguments{
\item{sourcevar}{An input character vector of industry codes to look up.}

\item{origin}{A string indicating one of the following industry/product/patent classifications: "NAICS2002", "NAICS2007", "NAICS2012", "NAICS2017", "NAICS" (combined), "USPC2012", "USPC (combined)".}

\item{setting}{Choose one of the three available measures from Bilir (2014).
\itemize{
  \item{"MeanACL": The mean citation time for each patent class (USPC). This is the default measure. Larger values are associated with longer product lifecycles and larger imitation risks.}
  \item{"MeanP75CL": The 75th-percentile citation times for each patent class (USPC). This measurement is less susceptible to variation across sectors, due to "unimportant" patents.}
  \item{"MeanP85CL": The 85th-percentile citation times for each patent class (USPC). This measurement is less susceptible to variation across sectors, due to "unimportant" patents.}
}}
}
\value{
The mean time lapse between the cited-patent's grant date and a subsequetion citation, averaged at each patent class (USPC) and at each industry (NAICS).
}
\description{
Returns product lifecycle lengths by manufacturing industry, using forward citatin lags, based on Bronwyn et al. (2001) and Bilir (2014).
}
\examples{
# NAICS
get_lifecycle(sourcevar = c("31", "32", "33"), origin = "NAICS",
              setting = "MeanACL")

get_lifecycle(sourcevar = c("3111", "3331"), origin = "NAICS",
              setting = "MeanACL")

# USPC
get_lifecycle(sourcevar = c("002", "280"), origin = "USPC",
              setting = "MeanACL")

# USPC
get_lifecycle(sourcevar = c("002/102", "800/292"), origin = "USPC",
              setting = "MeanACL")
}
\references{
\itemize{
  \item{Hall, Bronwyn H., Adam B. Jaffe, and Manuel Trajtenberg. "The NBER patent citation data file: Lessons, insights and methodological tools." (2001).}
  \item{Bilir, L. Kamran. "Patent laws, product life-cycle lengths, and multinational activity." American Economic Review 104.7 (2014).}
}
}
