% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_upstream.R, R/get_upstream_test.R
\name{get_upstream}
\alias{get_upstream}
\title{Looking Up the Upstreamness and Downstreamness of Industries}
\source{
Data from Pol Antras' webpage <https://scholar.harvard.edu/antras/publications/measurement-upstreamness-and-downstreamness-global-valuechains>.

Data from Pol Antras' webpage <https://scholar.harvard.edu/antras/publications/measurement-upstreamness-and-downstreamness-global-valuechains>.
}
\usage{
get_upstream(
  sourcevar,
  origin,
  country,
  year,
  setting = "GVC_Ui",
  detailed = FALSE
)

get_upstream(
  sourcevar,
  origin,
  country,
  year,
  setting = "GVC_Ui",
  detailed = FALSE
)
}
\arguments{
\item{sourcevar}{An input character vector of industry codes to look up.}

\item{origin}{A string indicating one of the following industry/product classifications: "HS0" (1988/92), "HS1" (1996), "HS2" (2002), "HS3" (2007), "HS4" (2012), "HS5" (2017), "HS" (combined), "SITC1" (1950), "SITC2" (1974), "SITC3" (1985), "SITC4" (2006), "NAICS2002", "NAICS2007", "NAICS2012", "NAICS2017", "ISIC2" (1968), "ISIC3" (1989), "ISIC4" (2008).}

\item{country}{A string setting the ISO 3-letter country code for which to return values (default = "USA"). Antras and Chor (2018) provide estimates for 40 countries and the Rest of the World (RoW). For a list of available countries, load the package and type "unique(upstream$ISO3C)".}

\item{year}{An integer setting the year for which to return values. Antras and Chor (2018) provide estimates for 1995-2011. The default returns estimates for 2011.}

\item{setting}{Choose one of the four available measures from Antras and Chor (2018).
\itemize{
  \item{"GVC_Ui": Upstreamness (net inventories correction). This is the defult measure. Larger values are associated with higher levels of upstreamness.}
  \item{"GVC_FUGOi": Final-use to gross-output (net inventories correction). Lower values are associated with higher levels of upstreamness.}
  \item{"GVC_Di": Downstreamness (net inventories correction). Larger values are associated with higher levels of downstreamness.}
  \item{"GVC_VAGOi": Value-added to gross-output (net inventories correction). Lower values are associated with higher levels of downstreamness}
}}

\item{detailed}{Choose whether to use detailed industry-level GVC_Ui from Antras and Chor (2012).
\itemize{
  \item{"FALSE": Do not report detailed measures. This is the default.}
  \item{"TRUE": Report the detailed measures.}
}}
}
\value{
Concords each element of the input vector to 2-digit ISIC3 codes, then uses the corresponding codes as input to extract estimates of upstreamness or downstreamness.

Concords each element of the input vector to 2-digit ISIC3 codes, then uses the corresponding codes as input to extract estimates of upstreamness or downstreamness.
}
\description{
Returns Antras and Chor (2018)'s measures of industry upstreamness and downstreamness.

Returns Antras and Chor (2018)'s measures of industry upstreamness and downstreamness.
}
\examples{
# ISIC3
get_upstream(sourcevar = c("01", "29", "29", "80"), origin = "ISIC3",
             country = "USA", year = "2011",
             setting = "GVC_Ui")

# HS5
get_upstream(sourcevar = c("0101", "0301", "7014", "8420"), origin = "HS5",
             country = "USA", year = "2011",
             setting = "GVC_Ui")
# ISIC3
get_upstream(sourcevar = c("01", "29", "29", "80"), origin = "ISIC3",
             country = "USA", year = "2011",
             setting = "GVC_Ui", detailed = FALSE)

# HS5
get_upstream(sourcevar = c("0101", "0301", "7014", "8420"), origin = "HS5",
             country = "USA", year = "2012",
             setting = "GVC_Ui", detailed = TRUE)
}
\references{
Antras, Pol, and Davin Chor. 2018. "On the Measurement of Upstreamness and Downstreamness in Global Value Chains." World Trade Evolution: Growth, Productivity and Employment, 126-194. Taylor & Francis Group.

Antras, Pol, and Davin Chor. 2018. "On the Measurement of Upstreamness and Downstreamness in Global Value Chains." World Trade Evolution: Growth, Productivity and Employment, 126-194. Taylor & Francis Group.
}
