% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs0_desc}
\alias{hs0_desc}
\title{HS0 Description}
\format{A data frame with 6380 rows and 2 variables:
\describe{
  \item{code}{HS0 Code}
  \item{desc}{HS0 Description}
}}
\source{
\url{https://comtrade.un.org/data/cache/classificationH0.json}
}
\usage{
hs0_desc
}
\description{
A dataset containing the description of products under the HS0
classification.
}
\keyword{datasets}
