% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs0_isic3}
\alias{hs0_isic3}
\title{HS0-ISIC3 Concordance}
\format{
A data frame with 5018 rows and 7 variables:
\describe{
  \item{HS0_6d}{6-digit HS0 Code}
  \item{HS0_4d}{4-digit HS0 Code}
  \item{HS0_2d}{2-digit HS0 Code}
  \item{ISIC3_4d}{4-digit ISIC3 Code}
  \item{ISIC3_3d}{3-digit ISIC3 Code}
  \item{ISIC3_2d}{2-digit ISIC3 Code}
  \item{ISIC3_1d}{1-digit ISIC3 Code}
}
}
\source{
https://wits.worldbank.org/product_concordance.html
}
\usage{
hs0_isic3
}
\description{
A dataset containing concordances between HS0 and ISIC3 classifications.
}
\keyword{datasets}
