% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs0_naics}
\alias{hs0_naics}
\title{HS0-NAICS Concordance}
\format{
A data frame with 8058 rows and 6 variables:
\describe{
  \item{HS0_6d}{6-digit HS0 Code}
  \item{HS0_4d}{4-digit HS0 Code}
  \item{HS0_2d}{2-digit HS0 Code}
  \item{NAICS_6d}{6-digit NAICS Code}
  \item{NAICS_4d}{4-digit NAICS Code}
  \item{NAICS_2d}{2-digit NAICS Code}
}
}
\source{
Pierce and Schott (2009, 2018), \url{https://spinup-000d1a-wp-offload-media.s3.amazonaws.com/faculty/wp-content/uploads/sites/47/2019/06/hssicnaics_20181015.zip}
}
\usage{
hs0_naics
}
\description{
A dataset containing concordances between HS0 and NAICS (combined) classifications.
}
\keyword{datasets}
