% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs0_naics2017}
\alias{hs0_naics2017}
\title{HS0-NAICS2017 Concordance}
\format{
A data frame with 5691 rows and 6 variables:
\describe{
  \item{HS0_6d}{6-digit HS0 Code}
  \item{HS0_4d}{4-digit HS0 Code}
  \item{HS0_2d}{2-digit HS0 Code}
  \item{NAICS2017_6d}{6-digit NAICS2017 Code}
  \item{NAICS2017_4d}{4-digit NAICS2017 Code}
  \item{NAICS2017_2d}{2-digit NAICS2017 Code}
}
}
\source{
Pierce and Schott (2009, 2018), \url{https://spinup-000d1a-wp-offload-media.s3.amazonaws.com/faculty/wp-content/uploads/sites/47/2019/06/hssicnaics_20181015.zip}
}
\usage{
hs0_naics2017
}
\description{
A dataset containing concordances between HS0 and NAICS2017 classification.
}
\keyword{datasets}
