% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs0_sitc2}
\alias{hs0_sitc2}
\title{HS0-SITC2 Concordance}
\format{A data frame with 5017 rows and 8 variables:
\describe{
  \item{HS0_6d}{6-digit HS0 Code}
  \item{HS0_4d}{4-digit HS0 Code}
  \item{HS0_2d}{2-digit HS0 Code}
  \item{SITC2_5d}{5-digit SITC2 Code}
  \item{SITC2_4d}{4-digit SITC2 Code}
  \item{SITC2_3d}{3-digit SITC2 Code}
  \item{SITC2_2d}{2-digit SITC2 Code}
  \item{SITC2_1d}{1-digit SITC2 Code}
}}
\source{
\url{https://wits.worldbank.org/product_concordance.html}
}
\usage{
hs0_sitc2
}
\description{
A dataset containing concordances between HS0 and SITC2 classification.
The variables are as follows:
}
\keyword{datasets}
