% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs1_bec4}
\alias{hs1_bec4}
\title{HS1-BEC4 Concordance}
\format{
A data frame with 5,334 rows and 6 variables:
\describe{
  \item{HS1_6d}{6-digit HS1 Code}
  \item{HS1_4d}{4-digit HS1 Code}
  \item{HS1_2d}{2-digit HS1 Code}
  \item{BEC4_3d}{3-digit BEC4 Code}
  \item{BEC4_2d}{2-digit BEC4 Code}
  \item{BEC4_1d}{1-digit BEC4 Code}
}
}
\source{
https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp
}
\usage{
hs1_bec4
}
\description{
A dataset containing concordances between HS1 and BEC4 classifications.
}
\keyword{datasets}
