% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs1_desc}
\alias{hs1_desc}
\title{HS1 Description}
\format{A data frame with 6473 rows and 2 variables:
\describe{
  \item{code}{HS1 Code}
  \item{desc}{HS1 Description}
}}
\source{
\url{https://comtrade.un.org/data/cache/classificationH1.json}
}
\usage{
hs1_desc
}
\description{
A dataset containing the description of products under the HS1
classification. The variables are as follows:
}
\keyword{datasets}
