% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs1_sitc1}
\alias{hs1_sitc1}
\title{HS1-SITC1 Concordance}
\format{A data frame with 5106 rows and 8 variables:
\describe{
  \item{HS1_6d}{6-digit HS1 Code}
  \item{HS1_4d}{4-digit HS1 Code}
  \item{HS1_2d}{2-digit HS1 Code}
  \item{SITC1_5d}{5-digit SITC1 Code}
  \item{SITC1_4d}{4-digit SITC1 Code}
  \item{SITC1_3d}{3-digit SITC1 Code}
  \item{SITC1_2d}{2-digit SITC1 Code}
  \item{SITC1_1d}{1-digit SITC1 Code}
}}
\source{
\url{https://wits.worldbank.org/product_concordance.html}
}
\usage{
hs1_sitc1
}
\description{
A dataset containing concordances between HS1 and SITC1 classification.
The variables are as follows:
}
\keyword{datasets}
