% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs1_sitc2}
\alias{hs1_sitc2}
\title{HS1-SITC2 Concordance}
\format{
A data frame with 5111 rows and 8 variables:
\describe{
  \item{HS1_6d}{6-digit HS1 Code}
  \item{HS1_4d}{4-digit HS1 Code}
  \item{HS1_2d}{2-digit HS1 Code}
  \item{SITC2_5d}{5-digit SITC2 Code}
  \item{SITC2_4d}{4-digit SITC2 Code}
  \item{SITC2_3d}{3-digit SITC2 Code}
  \item{SITC2_2d}{2-digit SITC2 Code}
  \item{SITC2_1d}{1-digit SITC2 Code}
}
}
\source{
https://wits.worldbank.org/product_concordance.html
}
\usage{
hs1_sitc2
}
\description{
A dataset containing concordances between HS1 and SITC2 classification.
}
\keyword{datasets}
