% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs1_sitc3}
\alias{hs1_sitc3}
\title{HS1-SITC3 Concordance}
\format{
A data frame with 5111 rows and 8 variables:
\describe{
  \item{HS1_6d}{6-digit HS1 Code}
  \item{HS1_4d}{4-digit HS1 Code}
  \item{HS1_2d}{2-digit HS1 Code}
  \item{SITC3_5d}{5-digit SITC3 Code}
  \item{SITC3_4d}{4-digit SITC3 Code}
  \item{SITC3_3d}{3-digit SITC3 Code}
  \item{SITC3_2d}{2-digit SITC3 Code}
  \item{SITC3_1d}{1-digit SITC3 Code}
}
}
\source{
\url{https://wits.worldbank.org/product_concordance.html}
}
\usage{
hs1_sitc3
}
\description{
A dataset containing concordances between HS1 and SITC3 classification.
}
\keyword{datasets}
