% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs1_sitc4}
\alias{hs1_sitc4}
\title{HS1-SITC4 Concordance}
\format{A data frame with 5111 rows and 8 variables:
\describe{
  \item{HS1_6d}{6-digit HS1 Code}
  \item{HS1_4d}{4-digit HS1 Code}
  \item{HS1_2d}{2-digit HS1 Code}
  \item{SITC4_5d}{5-digit SITC4 Code}
  \item{SITC4_4d}{4-digit SITC4 Code}
  \item{SITC4_3d}{3-digit SITC4 Code}
  \item{SITC4_2d}{2-digit SITC4 Code}
  \item{SITC4_1d}{1-digit SITC4 Code}
}}
\source{
\url{https://wits.worldbank.org/product_concordance.html}
}
\usage{
hs1_sitc4
}
\description{
A dataset containing concordances between HS1 and SITC4 classification.
}
\keyword{datasets}
