% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs2_desc}
\alias{hs2_desc}
\title{HS2 Description}
\format{
A data frame with 6568 rows and 2 variables:
\describe{
  \item{code}{HS2 Code}
  \item{desc}{HS2 Description}
}
}
\source{
\url{https://comtrade.un.org/data/cache/classificationH2.json}
}
\usage{
hs2_desc
}
\description{
A dataset containing the description of products under the HS2
classification.
}
\keyword{datasets}
