% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs2_hs0}
\alias{hs2_hs0}
\title{HS2-HS0 Concordance}
\format{
A data frame with 5223 rows and 6 variables:
\describe{
  \item{HS2_6d}{6-digit HS2 Code}
  \item{HS2_4d}{4-digit HS2 Code}
  \item{HS2_2d}{2-digit HS2 Code}
  \item{HS0_6d}{6-digit HS0 Code}
  \item{HS0_4d}{4-digit HS0 Code}
  \item{HS0_2d}{2-digit HS0 Code}
}
}
\source{
\url{https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp}
}
\usage{
hs2_hs0
}
\description{
A dataset containing concordances between HS2 and HS0 classification.
}
\keyword{datasets}
