% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs2_hs1}
\alias{hs2_hs1}
\title{HS2-HS1 Concordance}
\format{
A data frame with 5223 rows and 6 variables:
\describe{
  \item{HS2_6d}{6-digit HS2 Code}
  \item{HS2_4d}{4-digit HS2 Code}
  \item{HS2_2d}{2-digit HS2 Code}
  \item{HS1_6d}{6-digit HS1 Code}
  \item{HS1_4d}{4-digit HS1 Code}
  \item{HS1_2d}{2-digit HS1 Code}
}
}
\source{
\url{https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp}
}
\usage{
hs2_hs1
}
\description{
A dataset containing concordances between HS2 and HS1 classification.
}
\keyword{datasets}
