% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs2_isic2}
\alias{hs2_isic2}
\title{HS2-ISIC2 Concordance}
\format{
A data frame with 5224 rows and 7 variables:
\describe{
  \item{HS2_6d}{6-digit HS2 Code}
  \item{HS2_4d}{4-digit HS2 Code}
  \item{HS2_2d}{2-digit HS2 Code}
  \item{ISIC2_4d}{4-digit ISIC2 Code}
  \item{ISIC2_3d}{3-digit ISIC2 Code}
  \item{ISIC2_2d}{2-digit ISIC2 Code}
  \item{ISIC2_1d}{1-digit ISIC2 Code}
}
}
\source{
https://wits.worldbank.org/product_concordance.html
}
\usage{
hs2_isic2
}
\description{
A dataset containing concordances between HS2 and ISIC2 classifications.
}
\keyword{datasets}
