% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs2_sitc1}
\alias{hs2_sitc1}
\title{HS2-SITC1 Concordance}
\format{
A data frame with 5217 rows and 8 variables:
\describe{
  \item{HS2_6d}{6-digit HS2 Code}
  \item{HS2_4d}{4-digit HS2 Code}
  \item{HS2_2d}{2-digit HS2 Code}
  \item{SITC1_5d}{5-digit SITC1 Code}
  \item{SITC1_4d}{4-digit SITC1 Code}
  \item{SITC1_3d}{3-digit SITC1 Code}
  \item{SITC1_2d}{2-digit SITC1 Code}
  \item{SITC1_1d}{1-digit SITC1 Code}
}
}
\source{
https://wits.worldbank.org/product_concordance.html
}
\usage{
hs2_sitc1
}
\description{
A dataset containing concordances between HS2 and SITC1 classification.
}
\keyword{datasets}
