% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs2_sitc2}
\alias{hs2_sitc2}
\title{HS2-SITC2 Concordance}
\format{
A data frame with 5222 rows and 8 variables:
\describe{
  \item{HS2_6d}{6-digit HS2 Code}
  \item{HS2_4d}{4-digit HS2 Code}
  \item{HS2_2d}{2-digit HS2 Code}
  \item{SITC2_5d}{5-digit SITC2 Code}
  \item{SITC2_4d}{4-digit SITC2 Code}
  \item{SITC2_3d}{3-digit SITC2 Code}
  \item{SITC2_2d}{2-digit SITC2 Code}
  \item{SITC2_1d}{1-digit SITC2 Code}
}
}
\source{
https://wits.worldbank.org/product_concordance.html
}
\usage{
hs2_sitc2
}
\description{
A dataset containing concordances between HS2 and SITC2 classifications.
}
\keyword{datasets}
