\name{hs2sitc}
\alias{hs2sitc}

\docType{data}
\title{HS-SITC2 conversion table}

\description{Quick conversion table used in elasticity look-up functions.}

\usage{data("rauch")}

\format{
  A data frame with 11542 observations on the following 2 variables.
  \itemize{
    \item{\code{hs}}
    \item{\code{sitc2}}
  }
}

%%\details{}
%%\examples{}

\source{Data from Jon Haveman's International Trade Data page: http://www.macalester.edu/research/economics/PAGE/HAVEMAN/Trade.Resources/TradeData.html#Rauch}

\keyword{ datasets, rauch }