% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs3_desc}
\alias{hs3_desc}
\title{HS3 Description}
\format{
A data frame with 6372 rows and 2 variables:
\describe{
  \item{code}{HS3 Code}
  \item{desc}{HS3 Description}
}
}
\source{
\url{https://comtrade.un.org/data/cache/classificationH3.json}
}
\usage{
hs3_desc
}
\description{
A dataset containing the description of products under the HS3
classification.
}
\keyword{datasets}
