% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs3_hs1}
\alias{hs3_hs1}
\title{HS3-HS1 Concordance}
\format{
A data frame with 5052 rows and 6 variables:
\describe{
  \item{HS3_6d}{6-digit HS3 Code}
  \item{HS3_4d}{4-digit HS3 Code}
  \item{HS3_2d}{2-digit HS3 Code}
  \item{HS1_6d}{6-digit HS1 Code}
  \item{HS1_4d}{4-digit HS1 Code}
  \item{HS1_2d}{2-digit HS1 Code}
}
}
\source{
\url{https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp}
}
\usage{
hs3_hs1
}
\description{
A dataset containing concordances between HS3 and HS1 classification.
}
\keyword{datasets}
