% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs3_hs2}
\alias{hs3_hs2}
\title{HS3-HS2 Concordance}
\format{A data frame with 5052 rows and 6 variables:
\describe{
  \item{HS3_6d}{6-digit HS3 Code}
  \item{HS3_4d}{4-digit HS3 Code}
  \item{HS3_2d}{2-digit HS3 Code}
  \item{HS2_6d}{6-digit HS2 Code}
  \item{HS2_4d}{4-digit HS2 Code}
  \item{HS2_2d}{2-digit HS2 Code}
}}
\source{
\url{https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp}
}
\usage{
hs3_hs2
}
\description{
A dataset containing concordances between HS3 and HS2 classification.
}
\keyword{datasets}
