% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs3_isic3}
\alias{hs3_isic3}
\title{HS3-ISIC3 Concordance}
\format{
A data frame with 5052 rows and 7 variables:
\describe{
  \item{HS3_6d}{6-digit HS3 Code}
  \item{HS3_4d}{4-digit HS3 Code}
  \item{HS3_2d}{2-digit HS3 Code}
  \item{ISIC3_4d}{4-digit ISIC3 Code}
  \item{ISIC3_3d}{3-digit ISIC3 Code}
  \item{ISIC3_2d}{2-digit ISIC3 Code}
  \item{ISIC3_1d}{1-digit ISIC3 Code}
}
}
\source{
https://wits.worldbank.org/product_concordance.html
}
\usage{
hs3_isic3
}
\description{
A dataset containing concordances between HS3 and ISIC3 classifications.
}
\keyword{datasets}
