% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs3_naics2012}
\alias{hs3_naics2012}
\title{HS3-NAICS2012 Concordance}
\format{
A data frame with 6269 rows and 6 variables:
\describe{
  \item{HS3_6d}{6-digit HS3 Code}
  \item{HS3_4d}{4-digit HS3 Code}
  \item{HS3_2d}{2-digit HS3 Code}
  \item{NAICS2012_6d}{6-digit NAICS2012 Code}
  \item{NAICS2012_4d}{4-digit NAICS2012 Code}
  \item{NAICS2012_2d}{2-digit NAICS2012 Code}
}
}
\source{
Pierce and Schott (2009, 2018), \url{https://spinup-000d1a-wp-offload-media.s3.amazonaws.com/faculty/wp-content/uploads/sites/47/2019/06/hssicnaics_20181015.zip}
}
\usage{
hs3_naics2012
}
\description{
A dataset containing concordances between HS3 and NAICS2012 classification.
}
\keyword{datasets}
