% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs3_sitc1}
\alias{hs3_sitc1}
\title{HS3-SITC1 Concordance}
\format{A data frame with 5045 rows and 8 variables:
\describe{
  \item{HS3_6d}{6-digit HS3 Code}
  \item{HS3_4d}{4-digit HS3 Code}
  \item{HS3_2d}{2-digit HS3 Code}
  \item{SITC1_5d}{5-digit SITC1 Code}
  \item{SITC1_4d}{4-digit SITC1 Code}
  \item{SITC1_3d}{3-digit SITC1 Code}
  \item{SITC1_2d}{2-digit SITC1 Code}
  \item{SITC1_1d}{1-digit SITC1 Code}
}}
\source{
\url{https://wits.worldbank.org/product_concordance.html}
}
\usage{
hs3_sitc1
}
\description{
A dataset containing concordances between HS3 and SITC1 classification.
}
\keyword{datasets}
