% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs3_sitc3}
\alias{hs3_sitc3}
\title{HS3-SITC3 Concordance}
\format{
A data frame with 5050 rows and 8 variables:
\describe{
  \item{HS3_6d}{6-digit HS3 Code}
  \item{HS3_4d}{4-digit HS3 Code}
  \item{HS3_2d}{2-digit HS3 Code}
  \item{SITC3_5d}{5-digit SITC3 Code}
  \item{SITC3_4d}{4-digit SITC3 Code}
  \item{SITC3_3d}{3-digit SITC3 Code}
  \item{SITC3_2d}{2-digit SITC3 Code}
  \item{SITC3_1d}{1-digit SITC3 Code}
}
}
\source{
\url{https://wits.worldbank.org/product_concordance.html}
}
\usage{
hs3_sitc3
}
\description{
A dataset containing concordances between HS3 and SITC3 classification.
}
\keyword{datasets}
