% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs3_sitc4}
\alias{hs3_sitc4}
\title{HS3-SITC4 Concordance}
\format{
A data frame with 5050 rows and 8 variables:
\describe{
  \item{HS3_6d}{6-digit HS3 Code}
  \item{HS3_4d}{4-digit HS3 Code}
  \item{HS3_2d}{2-digit HS3 Code}
  \item{SITC4_5d}{5-digit SITC4 Code}
  \item{SITC4_4d}{4-digit SITC4 Code}
  \item{SITC4_3d}{3-digit SITC4 Code}
  \item{SITC4_2d}{2-digit SITC4 Code}
  \item{SITC4_1d}{1-digit SITC4 Code}
}
}
\source{
\url{https://wits.worldbank.org/product_concordance.html}
}
\usage{
hs3_sitc4
}
\description{
A dataset containing concordances between HS3 and SITC4 classification.
}
\keyword{datasets}
