% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs4_desc}
\alias{hs4_desc}
\title{HS4 Description}
\format{A data frame with 6528 rows and 2 variables:
\describe{
  \item{code}{HS4 Code}
  \item{desc}{HS4 Description}
}}
\source{
\url{https://comtrade.un.org/data/cache/classificationH4.json}
}
\usage{
hs4_desc
}
\description{
A dataset containing the description of products under the HS4
classification.
}
\keyword{datasets}
