% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs4_sitc3}
\alias{hs4_sitc3}
\title{HS4-SITC3 Concordance}
\format{
A data frame with 5206 rows and 8 variables:
\describe{
  \item{HS4_6d}{6-digit HS4 Code}
  \item{HS4_4d}{4-digit HS4 Code}
  \item{HS4_2d}{2-digit HS4 Code}
  \item{SITC3_5d}{5-digit SITC3 Code}
  \item{SITC3_4d}{4-digit SITC3 Code}
  \item{SITC3_3d}{3-digit SITC3 Code}
  \item{SITC3_2d}{2-digit SITC3 Code}
  \item{SITC3_1d}{1-digit SITC3 Code}
}
}
\source{
\url{https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp}
}
\usage{
hs4_sitc3
}
\description{
A dataset containing concordances between HS4 and SITC3 classification.
}
\keyword{datasets}
