% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs4_sitc4}
\alias{hs4_sitc4}
\title{HS4-SITC4 Concordance}
\format{A data frame with 5205 rows and 8 variables:
\describe{
  \item{HS4_6d}{6-digit HS4 Code}
  \item{HS4_4d}{4-digit HS4 Code}
  \item{HS4_2d}{2-digit HS4 Code}
  \item{SITC4_5d}{5-digit SITC4 Code}
  \item{SITC4_4d}{4-digit SITC4 Code}
  \item{SITC4_3d}{3-digit SITC4 Code}
  \item{SITC4_2d}{2-digit SITC4 Code}
  \item{SITC4_1d}{1-digit SITC4 Code}
}}
\source{
\url{https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp}
}
\usage{
hs4_sitc4
}
\description{
A dataset containing concordances between HS4 and SITC4 classification.
}
\keyword{datasets}
