% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs5_bec4}
\alias{hs5_bec4}
\title{HS5-BEC4 Concordance}
\format{
A data frame with 5511 rows and 6 variables:
\describe{
  \item{HS5_6d}{6-digit HS5 Code}
  \item{HS5_4d}{4-digit HS5 Code}
  \item{HS5_2d}{2-digit HS5 Code}
  \item{BEC}{Original BEC4 Code}
  \item{BEC_2d}{2-digit BEC4 Code}
  \item{BEC_1d}{1-digit BEC4 Code}
}
}
\source{
https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp
}
\usage{
hs5_bec4
}
\description{
A dataset containing concordances between HS5 and BEC4 classifications.
}
\keyword{datasets}
