% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs5_desc}
\alias{hs5_desc}
\title{HS5 Description}
\format{A data frame with 6708 rows and 2 variables:
\describe{
  \item{code}{HS5 Code}
  \item{desc}{HS5 Description}
}}
\source{
\url{https://comtrade.un.org/data/cache/classificationH5.json}
}
\usage{
hs5_desc
}
\description{
A dataset containing the description of products under the HS5
classification.
}
\keyword{datasets}
