% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs5_hs3}
\alias{hs5_hs3}
\title{HS5-HS3 Concordance}
\format{
A data frame with 5388 rows and 6 variables:
\describe{
  \item{HS5_6d}{6-digit HS5 Code}
  \item{HS5_4d}{4-digit HS5 Code}
  \item{HS5_2d}{2-digit HS5 Code}
  \item{HS3_6d}{6-digit HS3 Code}
  \item{HS3_4d}{4-digit HS3 Code}
  \item{HS3_2d}{2-digit HS3 Code}
}
}
\source{
\url{https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp}
}
\usage{
hs5_hs3
}
\description{
A dataset containing concordances between HS5 and HS3 classification.
}
\keyword{datasets}
