% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs5_isic3}
\alias{hs5_isic3}
\title{HS5-ISIC3 Concordance}
\format{
A data frame with 5338 rows and 7 variables:
\describe{
  \item{HS5_6d}{6-digit HS5 Code}
  \item{HS5_4d}{4-digit HS5 Code}
  \item{HS5_2d}{2-digit HS5 Code}
  \item{ISIC3_4d}{4-digit ISIC3 Code}
  \item{ISIC3_3d}{3-digit ISIC3 Code}
  \item{ISIC3_2d}{2-digit ISIC3 Code}
  \item{ISIC3_1d}{1-digit ISIC3 Code}
}
}
\source{
https://wits.worldbank.org/product_concordance.html
}
\usage{
hs5_isic3
}
\description{
A dataset containing concordances between HS5 and ISIC3 classifications.
}
\keyword{datasets}
