% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs5_naics2002}
\alias{hs5_naics2002}
\title{HS5-NAICS2002 Concordance}
\format{
A data frame with 6795 rows and 6 variables:
\describe{
  \item{HS5_6d}{6-digit HS5 Code}
  \item{HS5_4d}{4-digit HS5 Code}
  \item{HS5_2d}{2-digit HS5 Code}
  \item{NAICS2002_6d}{6-digit NAICS2002 Code}
  \item{NAICS2002_4d}{4-digit NAICS2002 Code}
  \item{NAICS2002_2d}{2-digit NAICS2002 Code}
}
}
\source{
Pierce and Schott (2009, 2018), \url{https://spinup-000d1a-wp-offload-media.s3.amazonaws.com/faculty/wp-content/uploads/sites/47/2019/06/hssicnaics_20181015.zip}
}
\usage{
hs5_naics2002
}
\description{
A dataset containing concordances between HS5 and NAICS2002 classification.
}
\keyword{datasets}
