% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs5_sitc3}
\alias{hs5_sitc3}
\title{HS5-SITC3 Concordance}
\format{
A data frame with 5387 rows and 8 variables:
\describe{
  \item{HS5_6d}{6-digit HS5 Code}
  \item{HS5_4d}{4-digit HS5 Code}
  \item{HS5_2d}{2-digit HS5 Code}
  \item{SITC3_5d}{5-digit SITC3 Code}
  \item{SITC3_4d}{4-digit SITC3 Code}
  \item{SITC3_3d}{3-digit SITC3 Code}
  \item{SITC3_2d}{2-digit SITC3 Code}
  \item{SITC3_1d}{1-digit SITC3 Code}
}
}
\source{
\url{https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp}
}
\usage{
hs5_sitc3
}
\description{
A dataset containing concordances between HS5 and SITC3 classifications.
}
\keyword{datasets}
