% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs6_bec4}
\alias{hs6_bec4}
\title{HS6-BEC4 Concordance}
\format{
A data frame with 5927 rows and 6 variables:
\describe{
  \item{HS6_6d}{6-digit HS6 Code}
  \item{HS6_4d}{4-digit HS6 Code}
  \item{HS6_2d}{2-digit HS6 Code}
  \item{BEC4_3d}{3-digit BEC4 Code}
  \item{BEC4_2d}{2-digit BEC4 Code}
  \item{BEC4_1d}{1-digit BEC4 Code}
}
}
\source{
https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp
}
\usage{
hs6_bec4
}
\description{
A dataset containing concordances between HS6 and BEC4 classifications.
}
\keyword{datasets}
