% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs6_desc}
\alias{hs6_desc}
\title{HS6 Description}
\format{
A data frame with 6939 rows and 2 variables:
\describe{
  \item{code}{HS6 Code}
  \item{desc}{HS6 Description}
}
}
\source{
\url{https://comtrade.un.org/data/cache/classificationH6.json}
}
\usage{
hs6_desc
}
\description{
A dataset containing the description of products under the HS6
classification.
}
\keyword{datasets}
