% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs6_hs1}
\alias{hs6_hs1}
\title{HS6-HS1 Concordance}
\format{
A data frame with 5613 rows and 6 variables:
\describe{
  \item{HS6_6d}{6-digit HS6 Code}
  \item{HS6_4d}{4-digit HS6 Code}
  \item{HS6_2d}{2-digit HS6 Code}
  \item{HS1_6d}{6-digit HS1 Code}
  \item{HS1_4d}{4-digit HS1 Code}
  \item{HS1_2d}{2-digit HS1 Code}
}
}
\source{
\url{https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp}
}
\usage{
hs6_hs1
}
\description{
A dataset containing concordances between HS6 and HS1 classification.
}
\keyword{datasets}
