% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs6_isic2}
\alias{hs6_isic2}
\title{HS6-ISIC2 Concordance}
\format{
A data frame with 5613 rows and 7 variables:
\describe{
  \item{HS6_6d}{6-digit HS6 Code}
  \item{HS6_4d}{4-digit HS6 Code}
  \item{HS6_2d}{2-digit HS6 Code}
  \item{ISIC2_4d}{4-digit ISIC2 Code}
  \item{ISIC2_3d}{3-digit ISIC2 Code}
  \item{ISIC2_2d}{2-digit ISIC2 Code}
  \item{ISIC2_1d}{1-digit ISIC2 Code}
}
}
\source{
https://wits.worldbank.org/product_concordance.html
}
\usage{
hs6_isic2
}
\description{
A dataset containing concordances between HS6 and ISIC2 classifications.
}
\keyword{datasets}
