% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs6_naics}
\alias{hs6_naics}
\title{HS6-NAICS Concordance}
\format{
A data frame with 9451 rows and 8 variables:
\describe{
  \item{HS6_6d}{6-digit HS6 Code}
  \item{HS6_4d}{4-digit HS6 Code}
  \item{HS6_2d}{2-digit HS6 Code}
  \item{NAICS_6d}{6-digit NAICS Code}
  \item{NAICS_5d}{5-digit NAICS Code}
  \item{NAICS_4d}{4-digit NAICS Code}
  \item{NAICS_3d}{3-digit NAICS Code}
  \item{NAICS_2d}{2-digit NAICS Code}
}
}
\source{
Pierce and Schott (2009, 2018), \url{https://spinup-000d1a-wp-offload-media.s3.amazonaws.com/faculty/wp-content/uploads/sites/47/2019/06/hssicnaics_20181015.zip}
}
\usage{
hs6_naics
}
\description{
A dataset containing concordances between HS6 and NAICS (combined) classifications.
}
\keyword{datasets}
