% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_desc}
\alias{hs_desc}
\title{HS (Combined) Description}
\format{
A data frame with 8261 rows and 2 variables:
\describe{
  \item{code}{HS Code}
  \item{desc}{HS Description}
}
}
\usage{
hs_desc
}
\description{
A dataset containing the description of products under the HS (Combined)
classification (H0 to H6).
}
\keyword{datasets}
