% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_isic31}
\alias{hs_isic31}
\title{HS-ISIC3.1 Concordance}
\format{
A data frame with 6,873 rows and 7 variables:
\describe{
  \item{HS_6d}{6-digit HS Code}
  \item{HS_4d}{4-digit HS Code}
  \item{HS_2d}{2-digit HS Code}
  \item{ISIC3.1_4d}{4-digit ISIC3.1 Code}
  \item{ISIC3.1_3d}{3-digit ISIC3.1 Code}
  \item{ISIC3.1_2d}{2-digit ISIC3.1 Code}
  \item{ISIC3.1_1d}{1-digit ISIC3.1 Code}
}
}
\source{
https://wits.worldbank.org/product_concordance.html
}
\usage{
hs_isic31
}
\description{
A dataset containing concordances between HS (combined) and ISIC3.1 classifications.
}
\keyword{datasets}
