% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_isic4}
\alias{hs_isic4}
\title{HS-ISIC4 Concordance}
\format{
A data frame with 22,885 rows and 7 variables:
\describe{
  \item{HS_6d}{6-digit HS Code}
  \item{HS_4d}{4-digit HS Code}
  \item{HS_2d}{2-digit HS Code}
  \item{ISIC4_4d}{4-digit ISIC4 Code}
  \item{ISIC4_3d}{3-digit ISIC4 Code}
  \item{ISIC4_2d}{2-digit ISIC4 Code}
  \item{ISIC4_1d}{1-digit ISIC4 Code}
}
}
\source{
https://wits.worldbank.org/product_concordance.html
}
\usage{
hs_isic4
}
\description{
A dataset containing concordances between HS (combined) and ISIC4 classifications.
}
\keyword{datasets}
