% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_naics}
\alias{hs_naics}
\title{HS-NAICS Concordance}
\format{
A data frame with 10399 rows and 8 variables:
\describe{
  \item{HS_6d}{6-digit HS Code}
  \item{HS_4d}{4-digit HS Code}
  \item{HS_2d}{2-digit HS Code}
  \item{NAICS_6d}{6-digit NAICS Code}
  \item{NAICS_5d}{5-digit NAICS Code}
  \item{NAICS_4d}{4-digit NAICS Code}
  \item{NAICS_3d}{3-digit NAICS Code}
  \item{NAICS_2d}{2-digit NAICS Code}
}
}
\source{
Pierce and Schott (2009, 2018), \url{https://spinup-000d1a-wp-offload-media.s3.amazonaws.com/faculty/wp-content/uploads/sites/47/2019/06/hssicnaics_20181015.zip}
}
\usage{
hs_naics
}
\description{
A dataset containing concordances between HS (combined) and NAICS (combined) classifications.
}
\keyword{datasets}
