% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_sitc3}
\alias{hs_sitc3}
\title{HS-SITC3 Concordance}
\format{A data frame with 6826 rows and 8 variables:
\describe{
  \item{HS_6d}{6-digit HS Code}
  \item{HS_4d}{4-digit HS Code}
  \item{HS_2d}{2-digit HS Code}
  \item{SITC3_5d}{5-digit SITC3 Code}
  \item{SITC3_4d}{4-digit SITC3 Code}
  \item{SITC3_3d}{3-digit SITC3 Code}
  \item{SITC3_2d}{2-digit SITC3 Code}
  \item{SITC3_1d}{1-digit SITC3 Code}
}}
\source{
\url{https://wits.worldbank.org/product_concordance.html}; \url{https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp}
}
\usage{
hs_sitc3
}
\description{
A dataset containing concordances between HS (combined) and SITC3 classification.
}
\keyword{datasets}
