% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ipc2012_desc}
\alias{ipc2012_desc}
\title{IPC Description for year 2012}
\format{
A data frame with 71664 rows and 2 variables:
\describe{
  \item{code}{IPC Code}
  \item{desc}{IPC Description}
}
}
\source{
\url{https://www.wipo.int/classifications/ipc/en/ITsupport/Version20120101/index.html}
}
\usage{
ipc2012_desc
}
\description{
A dataset containing the description of utility patents under the WIPO classification in 2012
}
\keyword{datasets}
