% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isic2_desc}
\alias{isic2_desc}
\title{ISIC2 Description}
\format{A data frame with 276 rows and 2 variables:
\describe{
  \item{code}{ISIC2 Code}
  \item{desc}{ISIC2 Description}
}}
\source{
\url{https://unstats.un.org/unsd/classifications/Econ/ISIC}
}
\usage{
isic2_desc
}
\description{
A dataset containing the description of products under the ISIC2
classification.
}
\keyword{datasets}
