% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isic3.1_desc}
\alias{isic3.1_desc}
\title{ISIC3.1 Description}
\format{
A data frame with 540 rows and 2 variables:
\describe{
  \item{code}{ISIC3.1 Code}
  \item{desc}{ISIC3.1 Description}
}
}
\source{
\url{https://unstats.un.org/unsd/classifications/Econ/ISIC}
}
\usage{
isic3.1_desc
}
\description{
A dataset containing the description of products under the ISIC3.1
classification.
}
\keyword{datasets}
