% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isic31_isic2}
\alias{isic31_isic2}
\title{ISIC3.1-ISIC2 Concordance}
\format{
A data frame with 636 rows and 8 variables:
\describe{
  \item{ISIC3.1_4d}{4-digit ISIC3.1 Code}
  \item{ISIC3.1_3d}{3-digit ISIC3.1 Code}
  \item{ISIC3.1_2d}{2-digit ISIC3.1 Code}
  \item{ISIC3.1_1d}{1-digit ISIC3.1 Code}
  \item{ISIC2_4d}{4-digit ISIC2 Code}
  \item{ISIC2_3d}{3-digit ISIC2 Code}
  \item{ISIC2_2d}{2-digit ISIC2 Code}
  \item{ISIC2_1d}{1-digit ISIC2 Code}
}
}
\source{
\url{https://unstats.un.org/unsd/classifications/Econ/ISIC}
}
\usage{
isic31_isic2
}
\description{
A dataset containing concordances between ISIC3.1 and ISIC2 classifications.
}
\keyword{datasets}
