% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isic31_isic3}
\alias{isic31_isic3}
\title{ISIC3.1-ISIC3 Concordance}
\format{
A data frame with 316 rows and 8 variables:
\describe{
  \item{ISIC3.1_4d}{4-digit ISIC3.1 Code}
  \item{ISIC3.1_3d}{3-digit ISIC3.1 Code}
  \item{ISIC3.1_2d}{2-digit ISIC3.1 Code}
  \item{ISIC3.1_1d}{1-digit ISIC3.1 Code}
  \item{ISIC3_4d}{4-digit ISIC3 Code}
  \item{ISIC3_3d}{3-digit ISIC3 Code}
  \item{ISIC3_2d}{2-digit ISIC3 Code}
  \item{ISIC3_1d}{1-digit ISIC3 Code}
}
}
\source{
\url{https://unstats.un.org/unsd/classifications/Econ/ISIC}
}
\usage{
isic31_isic3
}
\description{
A dataset containing concordances between ISIC3.1 and ISIC3 classifications.
}
\keyword{datasets}
