% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isic3_desc}
\alias{isic3_desc}
\title{ISIC3 Description}
\format{
A data frame with 528 rows and 2 variables:
\describe{
  \item{code}{ISIC3 Code}
  \item{desc}{ISIC3 Description}
}
}
\source{
\url{https://unstats.un.org/unsd/classifications/Econ/ISIC}
}
\usage{
isic3_desc
}
\description{
A dataset containing the description of products under the ISIC3
classification.
}
\keyword{datasets}
