% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isic4_desc}
\alias{isic4_desc}
\title{ISIC4 Description}
\format{A data frame with 766 rows and 2 variables:
\describe{
  \item{code}{ISIC4 Code}
  \item{desc}{ISIC4 Description}
}}
\source{
\url{https://unstats.un.org/unsd/classifications/Econ/ISIC}
}
\usage{
isic4_desc
}
\description{
A dataset containing the description of products under the ISIC4
classification. The variables are as follows:
}
\keyword{datasets}
