% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isic4_isic2}
\alias{isic4_isic2}
\title{ISIC4-ISIC2 Concordance}
\format{
A data frame with 1706 rows and 8 variables:
\describe{
  \item{ISIC4_4d}{4-digit ISIC4 Code}
  \item{ISIC4_3d}{3-digit ISIC4 Code}
  \item{ISIC4_2d}{2-digit ISIC4 Code}
  \item{ISIC4_1d}{1-digit ISIC4 Code}
  \item{ISIC2_4d}{4-digit ISIC2 Code}
  \item{ISIC2_3d}{3-digit ISIC2 Code}
  \item{ISIC2_2d}{2-digit ISIC2 Code}
  \item{ISIC2_1d}{1-digit ISIC2 Code}
}
}
\source{
\url{https://unstats.un.org/unsd/classifications/Econ/ISIC}
}
\usage{
isic4_isic2
}
\description{
A dataset containing concordances between ISIC4 and ISIC2 classifications.
}
\keyword{datasets}
