% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isic4_isic3}
\alias{isic4_isic3}
\title{ISIC4-ISIC3 Concordance}
\format{
A data frame with 798 rows and 8 variables:
\describe{
  \item{ISIC4_4d}{4-digit ISIC4 Code}
  \item{ISIC4_3d}{3-digit ISIC4 Code}
  \item{ISIC4_2d}{2-digit ISIC4 Code}
  \item{ISIC4_1d}{1-digit ISIC4 Code}
  \item{ISIC3_4d}{4-digit ISIC3 Code}
  \item{ISIC3_3d}{3-digit ISIC3 Code}
  \item{ISIC3_2d}{2-digit ISIC3 Code}
  \item{ISIC3_1d}{1-digit ISIC3 Code}
}
}
\source{
\url{https://unstats.un.org/unsd/classifications/Econ/ISIC}
}
\usage{
isic4_isic3
}
\description{
A dataset containing concordances between ISIC4 and ISIC3 classifications.
}
\keyword{datasets}
