% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K_bassPCA.R
\name{K_bassPCA}
\alias{K_bassPCA}
\title{Estimate K matrix with bassPCA}
\usage{
K_bassPCA(
  modPCA,
  type = 1,
  prior = NULL,
  prior_func = NULL,
  mcmc.use = NULL,
  func.use = NULL
)
}
\arguments{
\item{modPCA}{a fitted model of class bassBasis from bassPCA function}

\item{type}{1 or 2. Use C_bassPCA or C_bassPCA_v2?}

\item{prior}{NULL (default) \verb{[0, 1]} prior for each variable. See details for required structure of prior}

\item{prior_func}{a vector of weights to use when summing over functional variable. Should be same length as func.use.}

\item{mcmc.use}{vector of mcmc indices to be used for both models. Otherwise, a matrix}

\item{func.use}{a vector of points of the functional variable to use}
}
\value{
A list returning the (posterior samples?) of the C matrix for each point specified in func.use
}
\description{
Closed form estimator of the K matrix using a BASS model
}
\details{
This function works by converting the linear combination of bass models to a single bass model. See C_bass for more details
}
