% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcbass2bass.R
\name{lcbass2bass}
\alias{lcbass2bass}
\title{Convert a linear combination of BASS models to a single BASS model}
\usage{
lcbass2bass(
  mod_list,
  weights = rep(1, length(mod_list)),
  yy = NULL,
  mcmc.use = NULL
)
}
\arguments{
\item{mod_list}{A list of bass models.}

\item{weights}{An optional vector of weights.}

\item{yy}{The data vector. Optional, but useful for some bass object methods.}

\item{mcmc.use}{set of indices telling which mcmc draws to use.}
}
\description{
A linear combination of BASS models is also a BASS model. This function takes a list of BASS models (all with the same data matrix xx.des)
and returns the resulting linear combination as a new BASS model.
One useful application of this function is to convert bassPCA to bass for a fixed time point.
Does not currently work for bass models with functional or categorical inputs.
}
\examples{
a <- 1
}
