% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lifecycle_NAICS2002_4d}
\alias{lifecycle_NAICS2002_4d}
\title{Product Life-cycle (NAICS, 4-digits)}
\format{
A data frame with 3 rows and 4 variables:
\describe{
  \item{source}{4-digit NAICS2002 Code}
  \item{MeanACL}{The mean of patent durations by patent class}
  \item{MeanP75CL}{The 75th percentile of patent durations by patent class}
  \item{MeanP85CL}{The 85th percentile of patent durations by patent class}
}
}
\source{
\url{https://www.aeaweb.org/articles?id=10.1257/aer.104.7.1979}
}
\usage{
lifecycle_NAICS2002_4d
}
\description{
A dataset containing patent citation lags, averaged across all patent classes pertaining to each manufacturing industry (NAICS, 4 digits) in 2002, based on Bilir (2014).
}
\references{
Bilir, L. Kamran. 2014. "Patent Laws, Product Life-Cycle Lengths, and Multinational Activity." American Economic Review, 104(7), 1979-2013.
}
\keyword{datasets}
