% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lifecycle_USPC2012_class}
\alias{lifecycle_USPC2012_class}
\title{Product Life-cycle (USPC, Class)}
\format{
A data frame with 3 rows and 4 variables:
\describe{
  \item{source}{Patent classification as class-level, including leading zeros}
  \item{MeanACL}{The mean of patent durations by patent classification}
  \item{MeanP75CL}{The 75th percentile of patent durations by patent classification}
  \item{MeanP85CL}{The 85th percentile of patent durations by patent classification}
}
}
\source{
\url{https://www.aeaweb.org/articles?id=10.1257/aer.104.7.1979}
}
\usage{
lifecycle_USPC2012_class
}
\description{
A dataset containing patent citation lags, averaged by patent class in 2012, based on Bilir (2014).
}
\references{
Bilir, L. Kamran. 2014. "Patent Laws, Product Life-Cycle Lengths, and Multinational Activity." American Economic Review, 104(7), 1979-2013.
}
\keyword{datasets}
