\name{long_codes}
\alias{long_codes}

\docType{data}
\title{List of separately concorded codes}

\description{
List containing names of codes listed in concord_long instead of concord_data.
}

%%  \usage{data("long_codes")}

\format{
  A list with 2 entries at present.
}

%%\details{}

\source{
Data consolidated from
- World Integrated Trade Solution (WITS), World Bank (http://wits.worldbank.org/product_concordance.html)
- UN Statistics Division (http://unstats.un.org/unsd/trade/conversions/HS%20Correlation%20and%20Conversion%20tables.htm)
- Schott and Pierce's site (http://faculty.som.yale.edu/peterschott/sub_international.htm)
}

\references{See also Justin R. Pierce and Peter K. Schott, "Concording U.S. Harmonized System Codes Over Time".}

\keyword{ datasets }
