% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_functions.R
\name{modified_borehole}
\alias{modified_borehole}
\title{A Modified Borehole Function}
\usage{
modified_borehole(xx, design = 0.5)
}
\arguments{
\item{xx}{5 inputs, restricted to the unit interval. More inputs can be used but they are completely inert.}

\item{design}{the radius of the borehole (typically rw)}
}
\value{
The output of the borehole function
}
\description{
This function is for testing, it is a modified borehole function designed to have a more interesting active subspace
}
\details{
PARAMETER RANGES
rw in \verb{[0.05, 0.15]} 	radius of borehole (m)
r in \verb{[100, 50000]} 	radius of influence (m)
Tu in \verb{[63070, 115600]}    	transmissivity of upper aquifer (m2/yr)
Hu in \verb{[990, 1110]} 	potentiometric head of upper aquifer (m)
Tl in \verb{[63.1, 116]} 	transmissivity of lower aquifer (m2/yr)
Hl in \verb{[700, 820]} 	potentiometric head of lower aquifer (m)
L in \verb{[1120, 1680]} 	length of borehole (m)
Kw in \verb{[9855, 12045]} 	hydraulic conductivity of borehole (m/yr)
}
