% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naics2002_desc}
\alias{naics2002_desc}
\title{NAICS 2002 Description}
\format{A data frame with 2341 rows and 2 variables:
\describe{
  \item{code}{NAICS 2002 Code}
  \item{desc}{NAICS 2002 Description}
}}
\source{
\url{https://www.census.gov/eos/www/naics/reference_files_tools/2002/naics_6_02.txt}
}
\usage{
naics2002_desc
}
\description{
A dataset containing the description of products under the NAICS 2002
classification.
}
\keyword{datasets}
