% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naics2002_isic31}
\alias{naics2002_isic31}
\title{NAICS2002-ISIC3.1 Concordance}
\format{
A data frame with 1960 rows and 7 variables:
\describe{
  \item{NAICS2002_6d}{6-digit NAICS2002 Code}
  \item{NAICS2002_4d}{4-digit NAICS2002 Code}
  \item{NAICS2002_2d}{2-digit NAICS2002 Code}
  \item{ISIC3.1_4d}{4-digit ISIC3.1 Code}
  \item{ISIC3.1_3d}{3-digit ISIC3.1 Code}
  \item{ISIC3.1_2d}{2-digit ISIC3.1 Code}
  \item{ISIC3.1_1d}{1-digit ISIC3.1 Code}
}
}
\source{
\url{https://www.census.gov/eos/www/naics/concordances/concordances.html}
}
\usage{
naics2002_isic31
}
\description{
A dataset containing concordances between NAICS2002 and ISIC3.1 classifications.
}
\keyword{datasets}
