% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naics2012_desc}
\alias{naics2012_desc}
\title{NAICS 2012 Description}
\format{
A data frame with 2229 rows and 2 variables:
\describe{
  \item{code}{NAICS 2012 Code}
  \item{desc}{NAICS 2012 Description}
}
}
\source{
\url{https://data.bls.gov/cew/apps/bls_naics/v2/bls_naics_app.htm#tab=download&naics=2012}
}
\usage{
naics2012_desc
}
\description{
A dataset containing the description of products under the NAICS 2012
classification.
}
\keyword{datasets}
