% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naics2012_naics2007}
\alias{naics2012_naics2007}
\title{NAICS2012-NAICS2007 Concordance}
\format{
A data frame with 1184 rows and 6 variables:
\describe{
  \item{NAICS2012_6d}{6-digit NAICS2012 Code}
  \item{NAICS2012_4d}{4-digit NAICS2012 Code}
  \item{NAICS2012_2d}{2-digit NAICS2012 Code}
  \item{NAICS2007_6d}{6-digit NAICS2007 Code}
  \item{NAICS2007_4d}{4-digit NAICS2007 Code}
  \item{NAICS2007_2d}{2-digit NAICS2007 Code}
}
}
\source{
\url{https://www.census.gov/eos/www/naics/concordances/concordances.html}
}
\usage{
naics2012_naics2007
}
\description{
A dataset containing concordances between NAICS2012 and NAICS2007 classification.
}
\keyword{datasets}
