% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naics2017_naics2002}
\alias{naics2017_naics2002}
\title{NAICS2017-NAICS2002 Concordance}
\format{
A data frame with 1212 rows and 6 variables:
\describe{
  \item{NAICS2017_6d}{6-digit NAICS2017 Code}
  \item{NAICS2017_4d}{4-digit NAICS2017 Code}
  \item{NAICS2017_2d}{2-digit NAICS2017 Code}
  \item{NAICS2002_6d}{6-digit NAICS2002 Code}
  \item{NAICS2002_4d}{4-digit NAICS2002 Code}
  \item{NAICS2002_2d}{2-digit NAICS2002 Code}
}
}
\source{
\url{https://www.census.gov/eos/www/naics/concordances/concordances.html}
}
\usage{
naics2017_naics2002
}
\description{
A dataset containing concordances between NAICS2017 and NAICS2002 classification.
}
\keyword{datasets}
