% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance_r.R
\name{plot.ConcordanceAnalysis}
\alias{plot.ConcordanceAnalysis}
\title{Plotting Function for object of class ConcordanceAnalysis}
\usage{
\method{plot}{ConcordanceAnalysis}(x, ...)
}
\arguments{
\item{x}{object of class "ConcordanceAnalysis"}

\item{...}{arguments to be passed to individual plot functions}
}
\description{
Plotting Function for object of class ConcordanceAnalysis
}
