% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance_r.R
\name{plot_active_grad_k}
\alias{plot_active_grad_k}
\title{Plot components of the kth Principle Gradient}
\usage{
plot_active_grad_k(obj, k = 1, vnames = NULL, ...)
}
\arguments{
\item{obj}{object of class "ConcordanceAnalysis"}

\item{k}{which principle gradient is desired?}

\item{vnames}{optional vector of variable names}

\item{...}{additional arguments passed to barplot}
}
\description{
Plot components of the kth Principle Gradient
}
