% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance_r.R
\name{plot_contributions}
\alias{plot_contributions}
\title{Plot contributions}
\usage{
plot_contributions(obj, ...)
}
\arguments{
\item{obj}{object of class "ConcordanceAnalysis"}

\item{...}{additional arguments passed to barplot}
}
\description{
Plots the contributions pi_f, pi_g, and pi_fg
}
