% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance_r.R
\name{plot_sensitivities}
\alias{plot_sensitivities}
\title{Plot sensitivities}
\usage{
plot_sensitivities(obj, vnames = NULL, ...)
}
\arguments{
\item{obj}{object of class "ConcordanceAnalysis"}

\item{vnames}{optional vector of variable names}

\item{...}{additional arguments passed to barplot}
}
\description{
The sensitivity of variable j (wrt to f) is defined as
sum_{i=1}^n pi_f(i) * delta_f(j,i)
The definition is similar for g or for fg
}
