% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance_r.R
\name{print.ConcordanceAnalysis}
\alias{print.ConcordanceAnalysis}
\title{Summary and Print functions}
\usage{
\method{print}{ConcordanceAnalysis}(x, ...)
}
\arguments{
\item{x}{object of class "ConcordanceAnalysis"}

\item{...}{Additional arguments (ignored)}
}
\description{
Prints a summary for an object of class "ConcordanceAnalysis"
}
