\name{Rauch}
\alias{Rauch}

\docType{data}
\title{Table of Rauch classifications}

\description{Rauch classification look-up table using 4-digit SITC4 codes}

\usage{data("rauch")}

\format{
  A data frame with 1189 observations on the following 3 variables.
  \itemize{
    \item{\code{sitc4}}
    \item{\code{con}}
    \item{\code{lib}}
  }
}

%%\details{}
%%\examples{}

\source{Data from Jon Haveman's International Trade Data page: http://www.macalester.edu/research/economics/PAGE/HAVEMAN/Trade.Resources/TradeData.html#Rauch}

\keyword{ datasets, rauch }