% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_calculation.R
\name{sampleCalculationRange}
\alias{sampleCalculationRange}
\title{Sample Calculation and Plot of Power Versus Sample Size (Range)}
\usage{
sampleCalculationRange(
  lower_bound,
  upper_bound,
  step = 5,
  data.type = c("gut", "oral", "infant"),
  method = c("amp", "wgs"),
  nSim = 100,
  alpha = 0.05,
  nOTU = 500,
  diff.otu.pct = 0.1,
  diff.otu.direct = "balanced",
  diff.otu.mode = "abundant",
  user_specified_otu = NULL,
  covariate.type = "binary",
  grp.ratio = 1,
  covariate.eff.mean = 1,
  covariate.eff.sd = 0,
  confounder.type = "none",
  conf.cov.cor = 0.6,
  conf.diff.otu.pct = 0,
  conf.nondiff.otu.pct = 0.1,
  confounder.eff.mean = 0,
  confounder.eff.sd = 0,
  error.sd = 0,
  depth.mu = 10000,
  depth.theta = 5,
  depth.conf.factor = 0,
  cont.conf = NULL,
  epsilon = NULL
)
}
\arguments{
\item{lower_bound}{An integer specifying the minimum sample size to evaluate.}

\item{upper_bound}{An integer specifying the maximum sample size to evaluate.}

\item{step}{An integer specifying the increment in sample size. Default is 5.}

\item{data.type}{A character string specifying the data type. Must be one of \code{"gut"}, \code{"oral"}, or \code{"infant"}.}

\item{method}{A character string specifying the sequencing/estimation method. Must be one of \code{"amp"} or \code{"wgs"}.}

\item{nSim}{Number of simulation replicates for each sample size. Default is 100.}

\item{alpha}{Significance level for the PERMANOVA test. Default is 0.05.}

\item{nOTU}{Number of OTUs to simulate. Default is 500.}

\item{diff.otu.pct}{Proportion of OTUs that are differentially abundant. Default is 0.1.}

\item{diff.otu.direct}{Direction of differential OTUs. Either \code{"balanced"} or \code{"unbalanced"}. Default is \code{"balanced"}.}

\item{diff.otu.mode}{Mode of differential OTUs. Options include \code{"abundant"}, \code{"rare"}, \code{"mix"}, or \code{"user_specified"}. Default is \code{"abundant"}.}

\item{user_specified_otu}{Optional vector specifying the OTUs that are differentially abundant when \code{diff.otu.mode} is \code{"user_specified"}. Default is \code{NULL}.}

\item{covariate.type}{Type of covariate; either \code{"binary"} or \code{"continuous"}. Default is \code{"binary"}.}

\item{grp.ratio}{Group ratio for binary covariates. Default is 1.}

\item{covariate.eff.mean}{Mean effect size for the covariate. Default is 1.}

\item{covariate.eff.sd}{Standard deviation of the covariate effect. Default is 0.}

\item{confounder.type}{Type of confounding variable. Options are \code{"none"}, \code{"binary"}, \code{"continuous"}, or \code{"both"}. Default is \code{"none"}.}

\item{conf.cov.cor}{Correlation between the covariate and the confounder. Default is 0.6.}

\item{conf.diff.otu.pct}{Proportion of differential OTUs affected by the confounder. Default is 0.}

\item{conf.nondiff.otu.pct}{Proportion of non-differential OTUs affected by the confounder. Default is 0.1.}

\item{confounder.eff.mean}{Mean effect size for the confounder. Default is 0.}

\item{confounder.eff.sd}{Standard deviation of the confounder effect. Default is 0.}

\item{error.sd}{Standard deviation for the error term. Default is 0.}

\item{depth.mu}{Mean sequencing depth. Default is 10000.}

\item{depth.theta}{Theta parameter for the negative binomial distribution of sequencing depth. Default is 5.}

\item{depth.conf.factor}{Factor for the covariate effect on sequencing depth. Default is 0.}

\item{cont.conf}{A vector of continuous confounder values. If not provided, new values will be generated for each simulation using \code{rnorm(n)}.}

\item{epsilon}{A vector of random error values. If not provided, new values will be generated for each simulation using \code{rnorm(n)}.}
}
\value{
A data frame with columns \code{sample_size} and \code{mean_power} showing the estimated power at each sample size.
}
\description{
This function calculates simulation-based power for a range of sample sizes. The user supplies a
lower bound and an upper bound, and the function computes the mean power at every 5-sample increment
by calling the existing \code{powerCalculation} function. A plot is generated showing the relationship
between sample size and power, and a data frame with the detailed results is returned.
}
\examples{
\dontrun{
  # Evaluate power for sample sizes from 50 to 200 (in increments of 5) for gut data using amp parameters,
  # with 50 simulation replicates and 200 OTUs.
  res <- sampleCalculationRange(lower_bound = 50, upper_bound = 200,
                                data.type = "gut", method = "amp", nSim = 50, nOTU = 200)
  print(res)
}

}
