% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sigma_hs0}
\alias{sigma_hs0}
\title{Sigma Table (3-Digit HS0)}
\format{A data frame with 11293 rows and 4 variables:
\describe{
  \item{iso3c}{ISO 3-letter Country Code}
  \item{HS0_3d}{3-digit HS0 Code}
  \item{HS0_2d}{2-digit HS0 Code}
  \item{sigma}{Import Demand Elasticity}
}}
\source{
\url{http://www.columbia.edu/~dew35/TradeElasticities/TradeElasticities.html}
}
\usage{
sigma_hs0
}
\description{
A dataset containing import demand elasticities by HS0 3-digit codes from Broda and Weinstein (QJE, 2006) for 73 countries.
}
\references{
C. Broda and D. Weinstein, "Globalization and the Gains from Variety," Quarterly Journal of Economics, Volume 121, Issue 2 - May 2006.
}
\keyword{datasets}
