% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sigma_sitc3}
\alias{sigma_sitc3}
\title{Sigma Table (5-Digit SITC3)}
\format{A data frame with 2716 rows and 7 variables:
\describe{
  \item{iso3c}{ISO 3-letter Country Code}
  \item{SITC3_5d}{5-digit SITC3 Code}
  \item{SITC3_4d}{4-digit SITC3 Code}
  \item{SITC3_3d}{3-digit SITC3 Code}
  \item{SITC3_2d}{2-digit SITC3 Code}
  \item{SITC3_1d}{1-digit SITC3 Code}
  \item{sigma}{Import Demand Elasticity}
}}
\source{
\url{http://www.columbia.edu/~dew35/TradeElasticities/TradeElasticities.html}
}
\usage{
sigma_sitc3
}
\description{
A dataset containing import demand elasticities for the United States by 5-digit SITC3 codes from Broda and Weinstein (QJE, 2006).
}
\references{
C. Broda and D. Weinstein, "Globalization and the Gains from Variety," Quarterly Journal of Economics, Volume 121, Issue 2 - May 2006.
}
\keyword{datasets}
