% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitc1_bec4}
\alias{sitc1_bec4}
\title{SITC1-BEC4 Concordance}
\format{
A data frame with 1,787 rows and 8 variables:
\describe{
  \item{SITC1_5d}{5-digit SITC1 Code}
  \item{SITC1_4d}{4-digit SITC1 Code}
  \item{SITC1_3d}{3-digit SITC1 Code}
  \item{SITC1_2d}{2-digit SITC1 Code}
  \item{SITC1_1d}{1-digit SITC1 Code}
  \item{BEC4_3d}{3-digit BEC4 Code}
  \item{BEC4_2d}{2-digit BEC4 Code}
  \item{BEC4_1d}{1-digit BEC4 Code}
}
}
\source{
https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp
}
\usage{
sitc1_bec4
}
\description{
A dataset containing concordances between SITC1 and BEC4 classifications.
}
\keyword{datasets}
