% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitc1_desc}
\alias{sitc1_desc}
\title{SITC1 Description}
\format{
A data frame with 3024 rows and 2 variables:
\describe{
  \item{code}{SITC1 Code}
  \item{desc}{SITC1 Description}
}
}
\source{
\url{https://comtrade.un.org/data/cache/classificationS1.json}
}
\usage{
sitc1_desc
}
\description{
A dataset containing the description of products under the SITC1
classification.
}
\keyword{datasets}
