% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitc2_desc}
\alias{sitc2_desc}
\title{SITC2 Description}
\format{
A data frame with 3988 rows and 2 variables:
\describe{
  \item{code}{SITC2 Code}
  \item{desc}{SITC2 Description}
}
}
\source{
\url{https://comtrade.un.org/data/cache/classificationS2.json}
}
\usage{
sitc2_desc
}
\description{
A dataset containing the description of products under the SITC2
classification.
}
\keyword{datasets}
