% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitc2_isic2}
\alias{sitc2_isic2}
\title{SITC2-ISIC2 Concordance}
\format{
A data frame with 1687 rows and 9 variables:
\describe{
  \item{SITC2_5d}{5-digit SITC2 Code}
  \item{SITC2_4d}{4-digit SITC2 Code}
  \item{SITC2_3d}{3-digit SITC2 Code}
  \item{SITC2_2d}{2-digit SITC2 Code}
  \item{SITC2_1d}{1-digit SITC2 Code}
  \item{ISIC2_4d}{4-digit ISIC2 Code}
  \item{ISIC2_3d}{3-digit ISIC2 Code}
  \item{ISIC2_2d}{2-digit ISIC2 Code}
  \item{ISIC2_1d}{1-digit ISIC2 Code}
}
}
\source{
\url{https://wits.worldbank.org/product_concordance.html}
}
\usage{
sitc2_isic2
}
\description{
A dataset containing concordances between SITC2 and ISIC2 classifications.
}
\keyword{datasets}
