% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitc2_rauch}
\alias{sitc2_rauch}
\title{SITC2-Rauch Concordance}
\format{
A data frame with 1189 rows and 3 variables:
\describe{
  \item{SITC2}{SITC2 Code}
  \item{CON}{Conservative classification}
  \item{LIB}{Liberal classification}
}
}
\source{
Data from Jon Haveman's International Trade Data page: \url{http://www.macalester.edu/research/economics/PAGE/HAVEMAN/Trade.Resources/TradeData.html#Rauch}
}
\usage{
sitc2_rauch
}
\description{
A dataset containing concordances between SITC2 and Rauch's classification.
}
\keyword{datasets}
