% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitc3_desc}
\alias{sitc3_desc}
\title{SITC3 Description}
\format{A data frame with 5951 rows and 2 variables:
\describe{
  \item{code}{SITC3 Code}
  \item{desc}{SITC3 Description}
}}
\source{
\url{https://comtrade.un.org/data/cache/classificationS3.json}
}
\usage{
sitc3_desc
}
\description{
A dataset containing the description of products under the SITC3
classification. The variables are as follows:
}
\keyword{datasets}
