% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitc3_isic3}
\alias{sitc3_isic3}
\title{SITC3-ISIC3 Concordance}
\format{
A data frame with 3069 rows and 9 variables:
\describe{
  \item{SITC3_5d}{5-digit SITC3 Code}
  \item{SITC3_4d}{4-digit SITC3 Code}
  \item{SITC3_3d}{3-digit SITC3 Code}
  \item{SITC3_2d}{2-digit SITC3 Code}
  \item{SITC3_1d}{1-digit SITC3 Code}
  \item{ISIC3_4d}{4-digit ISIC3 Code}
  \item{ISIC3_3d}{3-digit ISIC3 Code}
  \item{ISIC3_2d}{2-digit ISIC3 Code}
  \item{ISIC3_1d}{1-digit ISIC3 Code}
}
}
\source{
\url{https://ec.europa.eu/eurostat}
}
\usage{
sitc3_isic3
}
\description{
A dataset containing concordances between SITC3 and ISIC3 classifications.
}
\keyword{datasets}
