% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitc3_naics}
\alias{sitc3_naics}
\title{SITC3-NAICS Concordance}
\format{
A data frame with 6024 rows and 10 variables:
\describe{
  \item{SITC3_5d}{5-digit SITC3 Code}
  \item{SITC3_4d}{4-digit SITC3 Code}
  \item{SITC3_3d}{3-digit SITC3 Code}
  \item{SITC3_2d}{2-digit SITC3 Code}
  \item{SITC3_1d}{1-digit SITC3 Code}
  \item{NAICS_6d}{6-digit NAICS Code}
  \item{NAICS_5d}{5-digit NAICS Code}
  \item{NAICS_4d}{4-digit NAICS Code}
  \item{NAICS_3d}{3-digit NAICS Code}
  \item{NAICS_2d}{2-digit NAICS Code}
}
}
\source{
https://wits.worldbank.org/product_concordance.html; \url{https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp}; Pierce and Schott (2009, 2018), \url{https://spinup-000d1a-wp-offload-media.s3.amazonaws.com/faculty/wp-content/uploads/sites/47/2019/06/hssicnaics_20181015.zip}
}
\usage{
sitc3_naics
}
\description{
A dataset containing concordances between SITC3 and NAICS (combined) classification via HS (combined).
}
\keyword{datasets}
