% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitc4_bec4}
\alias{sitc4_bec4}
\title{SITC4-BEC4 Concordance}
\format{
A data frame with 2,972 rows and 8 variables:
\describe{
  \item{SITC4_5d}{5-digit SITC4 Code}
  \item{SITC4_4d}{4-digit SITC4 Code}
  \item{SITC4_3d}{3-digit SITC4 Code}
  \item{SITC4_2d}{2-digit SITC4 Code}
  \item{SITC4_1d}{1-digit SITC4 Code}
  \item{BEC4_3d}{3-digit BEC4 Code}
  \item{BEC4_2d}{2-digit BEC4 Code}
  \item{BEC4_1d}{1-digit BEC4 Code}
}
}
\source{
https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp
}
\usage{
sitc4_bec4
}
\description{
A dataset containing concordances between SITC4 and BEC4 classifications.
}
\keyword{datasets}
