% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitc4_desc}
\alias{sitc4_desc}
\title{SITC4 Description}
\format{
A data frame with 7239 rows and 2 variables:
\describe{
  \item{code}{SITC4 Code}
  \item{desc}{SITC4 Description}
}
}
\source{
\url{https://comtrade.un.org/data/cache/classificationS4.json}
}
\usage{
sitc4_desc
}
\description{
A dataset containing the description of products under the SITC4
classification.
}
\keyword{datasets}
