% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitc4_naics}
\alias{sitc4_naics}
\title{SITC4-NAICS Concordance}
\format{
A data frame with 5714 rows and 8 variables:
\describe{
  \item{SITC4_5d}{5-digit SITC4 Code}
  \item{SITC4_4d}{4-digit SITC4 Code}
  \item{SITC4_3d}{3-digit SITC4 Code}
  \item{SITC4_2d}{2-digit SITC4 Code}
  \item{SITC4_1d}{1-digit SITC4 Code}
  \item{NAICS_6d}{6-digit NAICS Code}
  \item{NAICS_4d}{4-digit NAICS Code}
  \item{NAICS_2d}{2-digit NAICS Code}
}
}
\source{
\url{https://wits.worldbank.org/product_concordance.html}; \url{https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp}; Pierce and Schott (2009, 2018),\url{https://spinup-000d1a-wp-offload-media.s3.amazonaws.com/faculty/wp-content/uploads/sites/47/2019/06/hssicnaics_20181015.zip}
}
\usage{
sitc4_naics
}
\description{
A dataset containing concordances between SITC4 and NAICS (combined) classification via HS (combined).
}
\keyword{datasets}
