% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sitc4_sitc2}
\alias{sitc4_sitc2}
\title{SITC4-SITC2 Concordance}
\format{
A data frame with 3271 rows and 10 variables:
\describe{
  \item{SITC4_5d}{5-digit SITC4 Code}
  \item{SITC4_4d}{4-digit SITC4 Code}
  \item{SITC4_3d}{3-digit SITC4 Code}
  \item{SITC4_2d}{2-digit SITC4 Code}
  \item{SITC4_1d}{1-digit SITC4 Code}
  \item{SITC2_5d}{5-digit SITC2 Code}
  \item{SITC2_4d}{4-digit SITC2 Code}
  \item{SITC2_3d}{3-digit SITC2 Code}
  \item{SITC2_2d}{2-digit SITC2 Code}
  \item{SITC2_1d}{1-digit SITC2 Code}
}
}
\source{
\url{https://unstats.un.org/unsd/trade/classifications/correspondence-tables.asp}
}
\usage{
sitc4_sitc2
}
\description{
A dataset containing concordances between SITC4 and SITC2 classification.
}
\keyword{datasets}
