% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance_r.R
\name{summary.ConcordanceAnalysis}
\alias{summary.ConcordanceAnalysis}
\title{Summary and Print functions}
\usage{
\method{summary}{ConcordanceAnalysis}(object, ...)
}
\arguments{
\item{object}{object of class "ConcordanceAnalysis"}

\item{...}{Ignored}
}
\description{
Prints a summary for an object of class "ConcordanceAnalysis"
}
