% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upstream}
\alias{upstream}
\title{Upstreamness and Downstreamness Table}
\format{
A data frame with 24,395 rows and 7 variables:
\describe{
  \item{YEAR}{Year}
  \item{ISO3C}{ISO 3-letter Country Code}
  \item{WIOT2013_n}{WIOT 2013 Industry Code}
  \item{GVC_Ui}{Upstreamness (net inventories correction)}
  \item{GVC_FUGOi}{Final-use to gross-output (net inventories correction)}
  \item{GVC_Di}{Downstreamness (net inventories correction)}
  \item{GVC_VAGOi}{Value-added to gross-output (net inventories correction)}
}
}
\source{
\url{https://scholar.harvard.edu/antras/publications/measurement-upstreamness-and-downstreamness-global-valuechains}
}
\usage{
upstream
}
\description{
A dataset containing upstreamness and downstreamness estimates for 41 countries, 35 industries, and 17 years from Antras and Chor (2018).
}
\references{
Antras, Pol, and Davin Chor. 2018. "On the Measurement of Upstreamness and Downstreamness in Global Value Chains." World Trade Evolution: Growth, Productivity and Employment, 126-194. Taylor & Francis Group.
}
\keyword{datasets}
