% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{upstream_us_detailed}
\alias{upstream_us_detailed}
\title{US Detailed Upstreamness Table}
\format{
A data frame with 1,236 rows and 4 variables:
\describe{
  \item{YEAR}{Year}
  \item{CODE}{BEA Industry Code}
  \item{GVC_Ui}{Upstreamness (net inventories correction)}
  \item{BEA_CLASS}{Year of BEA Industry Code}
}
}
\source{
\url{https://scholar.harvard.edu/antras/publications/measuring-upstreamness-production-and-trade-flows}
}
\usage{
upstream_us_detailed
}
\description{
A dataset containing upstreamness estimates for 426 industries in 2002 and 405 industries in 2007 and 2012 based on Antras and Chor (2012).
}
\references{
Antras, Pol, and Davin Chor. 2012. "Measuring the Upstreamness of Production and Trade Flows." American Economic Review: Papers and Proceedings, Volume 102, Issue 3: 412-416.
}
\keyword{datasets}
