% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uspc2012_desc}
\alias{uspc2012_desc}
\title{USPC Description for year 2012}
\format{
A data frame with 164908 rows and 2 variables:
\describe{
  \item{code}{USPC Code}
  \item{desc}{USPC Description}
}
}
\source{
\url{https://www.uspto.gov/web/patents/classification/selectnumwithtitle.htm}
}
\usage{
uspc2012_desc
}
\description{
A dataset containing the description of utility patents under the USPC classification in 2012
}
\keyword{datasets}
