% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uspc2012_ipc2012}
\alias{uspc2012_ipc2012}
\title{USPC2012-IPC2012 Concordance}
\format{
A data frame with 258,216 rows and 7 variables:
\describe{
  \item{uspc_subclass}{USPC subclass}
  \item{uspc_class}{USPC class}
  \item{ipc_subgroup}{IPC subgroup}
  \item{ipc_group}{IPC group}
  \item{ipc_subclass}{IPC subclass}
  \item{ipc_class}{IPC class}
  \item{ipc_section}{IPC section}
}
}
\source{
https://www.uspto.gov/web/patents/classification/selectnumwithtitle.htm
}
\usage{
uspc2012_ipc2012
}
\description{
A dataset containing concordances between USPC2012 and IPC2012 classifications.
}
\keyword{datasets}
