% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uspc2012_naics2002}
\alias{uspc2012_naics2002}
\title{USPC2012-NAICS2002 Concordance}
\format{
A data frame with 862,586 rows and 5 variables:
\describe{
  \item{uspc_subclass}{USPC subclass}
  \item{uspc_class}{USPC class}
  \item{NAICS2002_4d}{4-digit NAICS2002 Code}
  \item{NAICS2002_3d}{3-digit NAICS2002 Code}
  \item{NAICS2002_2d}{2-digit NAICS2002 Code}
}
}
\source{
https://www.uspto.gov/web/offices/ac/ido/oeip/taf/naics/doc/naics_info.htm
}
\usage{
uspc2012_naics2002
}
\description{
A dataset containing concordances between USPC2012 and NAICS2002 classifications.
}
\keyword{datasets}
