% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiod_2013}
\alias{wiod_2013}
\title{ISIC3-WIOT2013 Concordance}
\format{
A data frame with 585 rows and 6 variables:
\describe{
  \item{ISIC3_4d}{4-digit ISIC3 Code}
  \item{ISIC3_3d}{3-digit ISIC3 Code}
  \item{ISIC3_2d}{2-digit ISIC3 Code}
  \item{ISIC3_1d}{1-digit ISIC3 Code}
  \item{WIOT2013_c}{WIOT 2013 Character Code}
  \item{WIOT2013_n}{WIOT 2013 Numeric Code}
}
}
\source{
\url{http://www.wiod.org/release13}
}
\usage{
wiod_2013
}
\description{
A dataset containing concordances between ISIC3 and WIOT2013 industry codes.
}
\keyword{datasets}
