% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cfg_Matrix.R
\name{Cfg_bass}
\alias{Cfg_bass}
\title{Estimate Cfg with BASS}
\usage{
Cfg_bass(mod1, mod2, prior = NULL, mcmc.use = NULL, scale01 = FALSE)
}
\arguments{
\item{mod1}{a fitted BASS model for first function}

\item{mod2}{a fitted BASS model for second function}

\item{prior}{NULL (default) \verb{[0, 1]} prior for each variable. See details for required structure of prior}

\item{mcmc.use}{vector of mcmc indices to be used for both models. Otherwise, a 2-column matrix with a pair of indices in each row.}

\item{scale01}{logical (default FALSE). When TRUE, the the C matix corresponds to the (0, 1)-scaled inputs rather than the original inputs.}
}
\value{
A list representing the posterior distribution of the Co-Constantine matrix (Cfg).
}
\description{
Closed form estimator of the Cfg matrix using a BASS model
}
\details{
prior should be a list of length p (one object for each variable). Each element of prior should be a named list with fields

dist - ("uniform", "normal").

trunc - truncation bounds (a, b)

mean - vector of means (mixture of normals only)

sigma - vector of sds (mixture of normals only)

weights - vector of mixture weights (mixture of normals only)
}
