% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_to_naics.R
\name{concord_hs_naics}
\alias{concord_hs_naics}
\title{Converting HS to NAICS Codes}
\usage{
concord_hs_naics(sourcevar, origin, destination, dest.digit = 6,
  all = FALSE)
}
\arguments{
\item{sourcevar}{An input character vector of HS codes. Allows 10, 6, 4, or 2 digits.}

\item{origin}{A string setting the input industry classification (HS).}

\item{destination}{A string setting the output industry classification (NAICS).}

\item{dest.digit}{An integer indicating the preferred number of digits for output codes. Allows 6, 4, or 2 digits. The default is 6 digits.}

\item{all}{Either TRUE or FALSE. If TRUE, the function will return (1) all matched outputs for each input, and (2) the share of occurrences for each matched output among all matched outputs. Users can use the shares as weights for more precise concordances. If FALSE, the function will only return the matched output with the largest share of occurrences (the mode match). If the mode consists of multiple matches, the function will return the first matched output.}
}
\value{
The matched output(s) for each element of the input vector. Either a list object when all = TRUE or a character vector when all = FALSE.
}
\description{
\code{concord_hs_naics} converts HS (combined) to NAICS (combined) codes
}
\examples{
# one input: one-to-one match
concord_hs_naics(sourcevar = "1206000069", origin = "HS", destination = "NAICS", all = FALSE)
concord_hs_naics(sourcevar = "1206000069", origin = "HS", destination = "NAICS", all = TRUE)

# two inputs: multiple-to-one match
concord_hs_naics(sourcevar = c("1206000069", "1206000061"), origin = "HS", destination = "NAICS", all = FALSE)
concord_hs_naics(sourcevar = c("1206000069", "1206000061"), origin = "HS", destination = "NAICS", all = TRUE)

# two inputs: repeated
concord_hs_naics(sourcevar = c("1206000069", "1206000069"), origin = "HS", destination = "NAICS", all = FALSE)

# one to multiple matches
concord_hs_naics(sourcevar = c("1206000069", "8546900000"), origin = "HS", destination = "NAICS", all = TRUE)

# sourcevar has different number of digits, gives an error
concord_hs_naics(sourcevar = c("1206000069", "120600006"), origin = "HS", destination = "NAICS", all = FALSE)

# if no match, will return NA
concord_hs_naics(sourcevar = c("1206000069", "1206000062"), origin = "HS", destination = "NAICS", all = FALSE)

# 4-digit inputs
concord_hs_naics(sourcevar = c("1206", "8546"), origin = "HS", destination = "NAICS", all = TRUE)

# 4-digit outputs
concord_hs_naics(sourcevar = c("1206000069", "1206000061"), origin = "HS", destination = "NAICS", dest.digit = 4, all = TRUE)
}
