\name{sitc2_rauch}
\alias{sitc2_rauch}

\docType{data}
\title{Table of Rauch classifications}

\description{
Rauch classification look-up table using 4-digit SITC2 codes
}

\usage{data("sitc2_rauch")}

\format{
  A data frame with 1189 observations on the following 3 variables.
  \itemize{
    \item{\code{SITC2}}
    \item{\code{CON}}
    \item{\code{LIB}}
  }
}

%%\details{}

\examples{
}

\source{
Data from
Jon Haveman's International Trade Data page: http://www.macalester.edu/research/economics/PAGE/HAVEMAN/Trade.Resources/TradeData.html#Rauch
}

\keyword{ datasets }
